Revert 43989 - This is a new version of the older patch. Main changes include: 1. Rebased to a never Git revision to make things easier to land. 2. Chrome will now preserve the command line switches except for those that are    blacklisted. 3. Fixed a race condition that would cause the browser to think it didn't exit    cleanly after it's been restarted. 4. Fixed minor nits and omissions (indentation, etc).
This patch adds a timer which fires every 6 hours and checks whether the
browser is in the "persistent" (background) mode, and whether there's an
update pending restart. If both conditions are true, the browser is
restarted with blacklisted command line keys and all loose values stripped. In order to
restart the browser in the background mode, the longlivedextensions
key is also added to the command line.

This change is Windowsonly, and it won't become fully functional until
Drew (atwilson) checks in his work that enables Chrome to go into background.

Review URL: http://codereview.chromium.org/1617001

TBR=atwilson@chromium.org
Review URL: http://codereview.chromium.org/1517027

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@44041 0039d316-1c4b-4281-b951-d872f2087c98