Fix comment in base/auto_reset.h to refer to AutoReset instead of AutoResetValue.
This inconsistency was introduced in r48644. The review history for that
change, http://codereview.chromium.org/2394001, indicates that AutoResetValue
was initially used in this file but not at any of the use sites. When it
changed to AutoReset in patch set 2, matching the use sites, the comment still
was not updated.

BUG=none, cleanup
TEST=none
Review URL: http://codereview.chromium.org/7253032

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@90707 0039d316-1c4b-4281-b951-d872f2087c98