Fix for the tools sanity unittest when ADDRESS_SANITIZER is defined on Windows.
The way we report an asan error is different on Windows, we use a RPC server to be able to get stack traces etc from the sandboxed processes. This make the tools sanity tests to fail on Windows. We should have some specific tests for Windows, but for now we should disable those tests on Windows.


BUG=


Review URL: https://chromiumcodereview.appspot.com/11753029

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@176705 0039d316-1c4b-4281-b951-d872f2087c98