base/time_posix.cc: Work around new integer overflow warning.
A new version of Clang warns about the previous version of this code:

  ../../base/time_posix.cc:212:58: warning: overflow in expression; result is 0 with type 'long' [-Winteger-overflow]
        milliseconds = std::numeric_limits<SysTime>::min() *
                                                           ^
  ../../base/time_posix.cc:215:59: warning: overflow in expression; result is -1000 with type 'long' [-Winteger-overflow]
        milliseconds = (std::numeric_limits<SysTime>::max() *
                                                            ^

The warning is legitimate: that code does overflow on LP64 platforms. However,
we will never hit that code on 64-bit platforms, because mktime() will not
overflow when time_t is a 64-bit type. I have added a CHECK about this and
worked around the warning by splitting up the expression.

BUG=24737
R=brettw@chromium.org, thakis@chromium.org

Review URL: https://codereview.chromium.org/15560002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@201735 0039d316-1c4b-4281-b951-d872f2087c98