Introduce a delay before a high priority binding is unbound.
This patch delays unbinding high priority connections for child services. This
makes sure that the renderers are not killed immediately after they are
unbound, which is neccessary to mitigate technical issues[1] with the way
renderers are being bound and strengthens the mechanism against possible
misbehavior of the system LRU process management.

[1] For instance, a content view can be detached for a quick moment while the
embedder displays a short-lived animation (e.g. opening a link in a background
tab) - we definitely don't want the underlying renderer to die during that
period.

BUG=259576
BUG=261339

Review URL: https://chromiumcodereview.appspot.com/19564005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@212445 0039d316-1c4b-4281-b951-d872f2087c98