Move a Thread's MessageLoop to the heap so it acts as a real object for leak reports.
We have a collection of leak reports where parts of a MessageLoop
aren't being destroyed. Since the MessageLoop is on the stack, these
all appear as separate reports and aren't clear about what's causing
them.

If moving the MessageLoop to the heap centralizes these reports, we'll
know that these threads are either terminating abruptly (pthread_exit,
pthread_cancel, or other unknown mechanisms, which we might be able to
detect with pthread_cleanup_push), or that the leak checkers are
failing to walk live threads at program exit.

BUG=269844,84224,16092

Review URL: https://chromiumcodereview.appspot.com/22707005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@218690 0039d316-1c4b-4281-b951-d872f2087c98