Reland "Add a HistogramRecorder class and use cases."
This CL adds a utility class to streamline the task of testing that metrics have been updated as expected. Specifically, the HistogramRecorder class allows a client to obtain the differential value of a given histogram in the interval since the given HistogramRecorder instance was created.

This reverts commit 3b2696c9a3d6eb662a2925003d7057893c83d8ba.

BUG=232414
TBR=rsleevi, mark, groby, thakis

Review URL: https://codereview.chromium.org/19866004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@242121 0039d316-1c4b-4281-b951-d872f2087c98