Make size of ashmem regions based on device's physical memory.
Ashmem regions created by the ashmem allocator used to be fixed to 32 MBytes
which can be a little small for high-end devices.

This change makes the allocator create ashmem regions of those sizes:
- 64 MBytes for 512 MBytes devices
- 128 MBytes for 1 GByte devices
- 256 MBytes for 2 GBytes...

This also adds an error code path in case ashmem creation fails which may 
happen if the address space is too fragmented. In this case the allocator
repetitively retries by dividing the ashmem region size by 2 until the ashmem
region reaches 32 MBytes (which is the minimum).

TBR=bulach@chromium.org

Review URL: https://codereview.chromium.org/119923002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@242497 0039d316-1c4b-4281-b951-d872f2087c98