Removing TryPostTask functionality
This is mostly a logical revert of
https://chromiumcodereview.appspot.com/14387002/

Because nobody uses it anymore. At that time there was
resistance about having this but there was a client (the audio code)
that wanted something like this to avoid polling. I don't know what
the audio code does (I so hope is not polling) but certainly is
not using this.

Also see
http://www.youtube.com/watch?v=BQ4yd2W50No

TBR=darin
BUG=none
TEST=none

Review URL: https://codereview.chromium.org/114333006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@242672 0039d316-1c4b-4281-b951-d872f2087c98