Add a StatisticsDeltaReader class and use cases
This CL adds a utility class to streamline the task of testing that
metrics have been updated as expected. Specifically, the
StatisticsDeltaReader class allows a client to obtain the differential
value of a given histogram in the interval since the given
StatisticsDeltaReader instance was created.

This reverts commit eb9c8a8e355cc2725fca6738a5959920ad80f556.

BUG=232414
TBR=rsleevy, groby

Review URL: https://codereview.chromium.org/120753002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@243303 0039d316-1c4b-4281-b951-d872f2087c98