Moved SuperFastHash function from base/hash.cc to base/third_party.
This code was written by Paul Hsieh under a 3-clause BSD license. It
belongs in third_party.

I did not move the existing implementation, but re-downloaded the code
from upstream:
http://www.azillionmonkeys.com/qed/hash.html
This version is formatted differently and is written in plain C, but it
is identical to the version that was previously in hash.cc.

Added LICENSE and README.chromium files, as well as adding license text
to the top of superfasthash.cc, as required by checklicenses.py.

BUG=347393

Review URL: https://codereview.chromium.org/181643004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@254437 0039d316-1c4b-4281-b951-d872f2087c98