Add optional public header checking to GN build
You can invoke this either using the "--check" argument to the gen command (for use on buildbots) or by running "gn check" (for developer on-demand use).

This adds support for "public" headers for a target, and an optional "--check" flag to the gen command that implements checkdeps-like include checking. Basically if you include a file that's declared in the build, it has to be in the public section of one of your dependents (or that dependent doesn't have a public section, which implies everything is public).

This roughly maps to Blaze's behavior around the public headers.

Moves modp_b64 build file to main tree.

Remove modp_b64 hack for older versions of VC missing stdint (this included basictypes which caused a header check failure).

I also added some base dependencies and some other minor build changes to solve some other issues identified by the check. The remaining one is that a file in base/metrics depends on ipc (!)

BUG=
R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/216903004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@262216 0039d316-1c4b-4281-b951-d872f2087c98