Pass the shared memory handle in to StatsTable rather than consulting the GlobaDescriptors.
Remove the IPC dependency from base and updates the DEPS files to prevent this from happening again (ipc is in the root dir's deps file so is implicitly allowed here).

The Chrome unit tests did a bunch of work to delete the stats file when running under Posix. But this filename is never used on Posix, so I deleted this code. The unit tests now use an anonymous stats table which this code now supports.

The stats table unit test is disabled. It does not work on Posix systems at all because handle sharing is not set up. I put a Windows ifdef around this code (leaving it disabled) so I could use the Windows-style initializers, and verified that the test passes on Windows if I undisable it.

R=jam@chromium.org

Review URL: https://codereview.chromium.org/224713017

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@262540 0039d316-1c4b-4281-b951-d872f2087c98