Add Skia to the GN build.
This makes the GN build of Skia compile and adds it to the GN build.
Previously, the configuration of the GN Skia tried to match the confusing array
of skia targets. This new version dispenses with that and just adds all files
(except SSE ones) to one target. I'm not even sure if it's necessary to split
out the SSE ones, but it seems nice since that target will get more complicated
when we add Arm and MIPS support.

I audited the defines again and added some to the main build config that had
been added to the GYP build recently.

This improves some operator error messages that I found confusing when I had errors in my file.

BUG=
R=djsollen@google.com, scottmg@chromium.org

Review URL: https://codereview.chromium.org/214883005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@262779 0039d316-1c4b-4281-b951-d872f2087c98