Re-land "Replaces use of ProcessInformationCopyDictionary ..."
This reverts commit 465bc8900e33ecc276b5f14499f4f9e8a01c700b.

Original CL was reverted because ProcessInfoRec::processAppSpec does not
exist on x64.  (Google search results claim it is named processAppRec instead.)

This patch changes the original CL by initializing the whole
ProcessInfoRect struct to 0 before use, rather than setting individual
members to NULL.

Original CL:
https://codereview.chromium.org/354243003/

Replaces use of ProcessInformationCopyDictionary with GetProcessInformation.

ProcessInformationCopyDictionary is crashing on 10.7.
This changes WasLaunchedAsLoginOrResumeItem to use lower-level calls
instead of ProcessInformationCopyDictionary, with the hope that it will
bypass the 10.7 crash. Other reports of crashes in this function on 10.7
claim to have been worked around, so there is reason to believe this will
not crash.

BUG=383029

Review URL: https://codereview.chromium.org/363503003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@280897 0039d316-1c4b-4281-b951-d872f2087c98