Handle bionic in safe_strerror_posix as well as glibc.
Android's bionic C library is intending to adopt the same semantics as
glibc for strerror_r: define the version that returns char* if the
source is compiled with _GNU_SOURCE instead of the POSIX version which
returns int. Add __BIONIC__ to the condition for
USE_HISTORICAL_STRERRO_R so that Chromium will still work.

BUG=

Review URL: https://codereview.chromium.org/491893002

Cr-Commit-Position: refs/heads/master@{#291050}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@291050 0039d316-1c4b-4281-b951-d872f2087c98