GN: Build //base with USE_SYMBOLIZE on Linux
In base.gyp, USE_SYMBOLIZE is added for Desktop Linux and
Chrome OS builds for both base and base_unittests. In the GN
build, it's added for 'is_linux' builds (which does not include
Android), but only for base_unittests. This adds the define for
//base as well.

BUG=335824
TEST=build base and base_unittests, make a debug build and force a DCHECK() -- you should get a useful stack trace
R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/583263003

Cr-Commit-Position: refs/heads/master@{#295885}