Remove build-time warning from pe_image.cc
This warning has been printed for over 1.5 years on every 64-bit build. Things
like this make people blind to warnings.

The code does look questionable on 64-bit, and the tests for this code are
apparently disabled for 64-bit too. This should be looked at (see bug), but
a warning on every build is not how this will happen.

BUG=167707,82385
R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/725293009

Cr-Commit-Position: refs/heads/master@{#305834}