mac: Implement custom AppKit Enter Fullscreen transition.
AppKit's default Fullscreen transition cross-fades between two static images: a
snapshot of the window before the transition began, and a snapshot of the
window resized to the full size of the screen. This has the problem that
frequently, when AppKit takes the snapshot of the final state of the window,
the web contents view has not yet drawn any content in the new size. The
animation itself looks weird, and then there's a flash when the animation
finishes where the snapshot of the old web content is replaced by the new web
content.

This custom fullscreen transition is very similar, except instead of taking a
snapshot of the window in its final state, the transition uses a live version
of the window in its final state. See the associated bug for more details about
the implementation.

BUG=414527

Review URL: https://codereview.chromium.org/846443004

Cr-Commit-Position: refs/heads/master@{#311717}