Add a base::TimeTicks::SnappedToNextTick method
Snapping a timestamp to the next phase (i.e. vsync) is a
common calculation we'll want to make in different parts
of the code.

Right now, it's only done in DelayBasedTimeSource.
However, this patch moves the calculation to base
in anticipation that it will be used in a couple
other places for latency calculations soon.

It's very easy to get wrong, given the number of bugs
we've fixed while the logic was in DelayBasedTimeSource,
so having a single shared implementation will avoid
duplicating mistakes.

BUG=360844

Review URL: https://codereview.chromium.org/913133006

Cr-Commit-Position: refs/heads/master@{#316093}