[tracing] memory maps dump provider skeleton
This CL introduces the skeleton (data model and dump provider stub)
for snapshotting memory maps into the traces (via MemoryDumpManager).
In the next CLs, the actual code to retrieve the data for the various
OSs (or at least some of them) will be introduced.

BUG=460884

Review URL: https://codereview.chromium.org/928723004

Cr-Commit-Position: refs/heads/master@{#317821}