Asynchronously pre-fault the native library pages.
Hard page faults are responsible for a significant fraction of startup
time on Android. This patch pre-touches the native library pages in a
low-priority background thread to warm up the page cache.

This leads to an improvement of up to 350ms out of 3400ms for "time to
navigation commit". See the linked bug for analysis, testing methodology
and experimental results.
This is only a first step, as it doesn't address the case where the
native library is loaded directly from the APK.

BUG=460438

Review URL: https://codereview.chromium.org/998643002

Cr-Commit-Position: refs/heads/master@{#320484}