Make swarming work w/ GN (kinda).
This patch implements the basic functionality needed to make swarming
and isolates work w/ GN. It relies on GN's existing functionality
for dumping the runtime dependencies needed for a target, and adds
a new command to MB that will take the runtime deps for a target,
compute the command line needed for the target, and then generate
the .isolate and .isolate.gen.json files needed for the
'isolate.py batcharchive' command.

We still need recipe-side work for swarming to actually work, however.

In addition, the way to manage the command lines for a target is a total
hack that will need to be cleaned up in subsequent patches.

R=maruel@chromium.org, brettw@chromium.org
BUG=480053

Review URL: https://codereview.chromium.org/1168513006

Cr-Commit-Position: refs/heads/master@{#333114}