Fix to set RenderThread priority correctly on Linux
On Linux, setting another thread's priority is not permitted
in renderer processes. The priority should be set as an option
on thread creations.

This change introduces ThreadPriority to SimpleThread::Options
as I did to Thread::Options so that RenderThread can use it
to set the thread priorities correctly.

BUG=468793

Review URL: https://codereview.chromium.org/1163053002

Cr-Commit-Position: refs/heads/master@{#333465}