Slow Reports - Add callbacks to histograms.
This CL adds the ability to set a callback to notify when the value is added to a histogram. These must be set in the process that the histogram will be modified in.

Ie. if foo.example1 is modified in the renderer, the call to StatisticsRecorder::SetCallback will need to be made in the renderer as well.

This will be used by slow reports to configure new triggers on UMA values.

BUG=504889

Review URL: https://codereview.chromium.org/1227163006

Cr-Commit-Position: refs/heads/master@{#338930}