Hide existing StrictMode violations.
While we work on to resolve these known StrictMode
violations, it's a poor experience for dogfooders to see
these every time they are triggered. It is still
important for developers and dogfooders to see and report
newly introduced violations with the StrictMode red border
flash.

BUG=508615,525781,525785,527415,527429

Review URL: https://codereview.chromium.org/1334533004

Cr-Commit-Position: refs/heads/master@{#348147}