Allow ParameterizedTests at the TestCase class level
A class-level ParameterizedTest is equivalent to having a ParameterizedTest
on each of the test methods. This is to save code lines if all tests of
the class need to be executed with certain parameters.

If a method has its own ParameterizedTest defined, then the test runs as if
there was a ParameterizedTest.Set defined on the method containing both
class-level ParameterizedTest and a method's own ParameterizedTest (that is,
two distinct test method invocations).

Class-level ParameterizedTests do not apply to subclasses.

BUG=526885

Review URL: https://codereview.chromium.org/1336133002

Cr-Commit-Position: refs/heads/master@{#350181}