Resize Breakpad guard region to the minimum required.
Current modern_linker creates a 16MB Breakpad guard region
ahead of the loaded library segments, to allow for packed
relocations. This 16MB is a conservative constant set to be
larger than the largest relocation packing. At address
reservation time the library's sizes are not yet know.

After the library has been loaded, the linker can uncover
the actual guard region size that is needed for packed
relocations, by obtaining the min vaddr of LOAD segments.

This change unmaps the excess portion of the initially
reserved Breakpad guard that is not needed. It frees up
around 10-14MB of address space for the process.

BUG=504410,499747

Review URL: https://codereview.chromium.org/1358373004

Cr-Commit-Position: refs/heads/master@{#350568}