views/mus: Fix showing menu in chrome in mus.
For menus in chrome-in-mus, we attempt to create a Widget hosted in a native
window. For desktop builds, this means we attempt to host in a HWND or an
X11 window created by DesktopNativeWidgetAura. But for mus, we want to use
NativeWidgetMus. To do this, use the native-widget factory first if it's
available.

BUG=none

Committed: https://crrev.com/4d6647c7c67c8fecbdc38dd7832f1e2ebdd74eab
Cr-Commit-Position: refs/heads/master@{#361151}

R=jam@chromium.org, sky@chromium.org

Review URL: https://codereview.chromium.org/1460003004 .

Cr-Commit-Position: refs/heads/master@{#361271}