Revert of Fix regression in memory-infra bots due to change in dump order (patchset #1 id:1 of https://codereview.chromium.org/1466293002/ )
Reason for revert:
Along with this CL, the memory-infra build was broken
(fixed in https://codereview.chromium.org/1474543003/). But, this
cause much noise in the metrics and did not give a clear idea if the
change fixed the original regression. So, I will revert this again
and re-land it to check if that actually makes a difference.

Original issue's description:
> Fix regression in memory-infra bots due to change in dump order
>
> crrev.com/1427963002 introduced a new memory-infra thread which runs
> after all dump providers are finished. This measures the process total
> memory usages. This was previously running at the start of a memory
> dump. This means that the memory used while taking memory dump of other
> providers will be included in the totals now and this metric has
> regressed. This CL is a temporary change that reverts this behaviour so
> that the perf bots are happy. Better fix for the problem will follow up
> soon.
>
> BUG=555584
>
> Committed: https://crrev.com/209e36ff91976a72d86a220c54d30de1b5c4425d
> Cr-Commit-Position: refs/heads/master@{#361126}

NOPRESUBMIT=true
NOTREECHECKS=true
BUG=555584

Review URL: https://codereview.chromium.org/1479003003

Cr-Commit-Position: refs/heads/master@{#361899}