Move ENABLE_PROFILING to build flags system.
This flag is only used in 3 files so having a global define is unnecessary.

Moves the build flag to compiler.gni so it can be included by the new target in base that generates the header, as well as the tcmalloc target.

The tcmalloc target now conditionally compiles the profiler depending on this flag. This is something the GYP build does but was inconsistent in GN. This results in the tcmalloc source files not being compiled in normal builds.

This should remove the static initializer in tcmalloc profiling in the normal build.

Independently, remove an iostream use in gin logging. Use printf instead.

BUG=559766
NOPRESUBMIT=true
(on account of changing the printf, which I did not add)

Review URL: https://codereview.chromium.org/1499773002

Cr-Commit-Position: refs/heads/master@{#363156}