Add more debugging CHECKS to diagnose an Android renderer crash.
Suspicion is that CreateTrialsFromSeed() is being called more than
once, so add some CHECKs that will help diagnose it. If it's not
being called twice but the same field trial is being registered
twice, this can also cause the same issue - so catch that too.

BUG=359406

Review URL: https://codereview.chromium.org/1632913004

Cr-Commit-Position: refs/heads/master@{#372252}