Enable attachment brokering of SharedMemoryHandle on Windows.
Whenever a SharedMemoryHandle associated with the current process is passed as a
parameter to a Chrome IPC message, the object is automatically brokered into the
destination process. This means it is no longer necessary to have a constructor
of SharedMemoryHandle that takes a ProcessHandle, since the SharedMemoryHandle
is always associated with the current process.

BUG=493414

Committed: https://crrev.com/7e0c1aa1e48a47ab8700bb57243a7202085aab0a
Cr-Commit-Position: refs/heads/master@{#370193}

Review URL: https://codereview.chromium.org/1588683003

Cr-Commit-Position: refs/heads/master@{#373174}