Add test for multithreaded ActiveVerifier behavior.
Previous to crrev.com/050e4b9 there was a race in the setting of the
g_active_verifier global that could have caused multiple instantiations.

This adds a test to verify the new behavior and detect any future
regressions. The test failing with the old code can be seen here:

http://pastebin.com/raw/iXz8j21C

BUG=592753
TEST=base_unittests --gtest_filter=ScopedHandleTest.MultiProcess

Review URL: https://codereview.chromium.org/1779333003

Cr-Commit-Position: refs/heads/master@{#380835}