Capture more details about failure to load ICU data file
The last round of debug stuff showed that
InitializeICUWithFileDescriptorInternal() is being called with an
invalid handle. This means that the File() ctor is failing. It's almost
definitely going to be something like access denied caused by AV, etc.,
but confirm that hypothesis by saving off the error on the file open.

R=jshin@chromium.org
BUG=445616

Review URL: https://codereview.chromium.org/1802713002

Cr-Commit-Position: refs/heads/master@{#381261}