Mac: Fix ScopedTypeRef constructor madness
The first part is to create a variant of ScopedCFTypeRef<IOSurfaceRef>
which hold a reference to the IOSurface and also marks it as "in use" (a
term with a very particular and poorly documented meaning, but it's
important in some places), and call it ScopedInUseIOSurface. Except that
makes everything crash.

The issue is that I pass a base::ScopedCFTypeRef<IOSurface> into the
gfx::ScopedInUseIOSurface's constructor. This "works" by casting the
argument into an IOSurfaceRef, and then handing that to the constructor,
which was then assuming ownership of the object, when it should be doing
a retain.

Add a constructor for ScopedTypeRef<A,TraitsX> that takes another
ScopedTypeRef<A,TraitsY>, which does the proper retain instead of
assume.

Add a move constructor while we're in the neighborhood.

TBR=scheib
BUG=

Review URL: https://codereview.chromium.org/1830043003

Cr-Commit-Position: refs/heads/master@{#383476}