Change SuperFastHash signature to use size_t, not int.
Move the size check and static_cast to from the generic Hash to the specific SuperFastHash wrapper.

Hash uses a reasonable size_t type, but the specific SuperFastHash uses an int for size. This is leaking a mistake from third_party code into our wrapper. Client code that wants to be clear that it is using SuperFastHash specifically (not just generic Hash) should be able to do so with a size_t type.

Review URL: https://codereview.chromium.org/1916193003

Cr-Commit-Position: refs/heads/master@{#389818}