Use ThreadTaskRunnerHandle instead of MessageLoop::current() in base/.
This CL replaces some uses of MessageLoop::current() with
ThreadTaskRunnerHandle in base/.

Why? The fact that there's a MessageLoop on the thread is an
unnecessary implementation detail. When browser threads are migrated to
base/task_scheduler, tasks will no longer have access to a MessageLoop
but they will be able to get the current ThreadTaskRunnerHandle.

BUG=616447

Review-Url: https://codereview.chromium.org/2061903002
Cr-Commit-Position: refs/heads/master@{#399711}