Add Browser Task Scheduler Flags
Adds the browser scheduler command line flags as well as exposing it
via chrome://flags.

Enabling the feature will place the browser in a forced variations
group, allowing it to use task scheduler parameters from there. If the
variations group is unavailable, then enabling this will still not
enable the browser task scheduler

BUG=553459,633389

Review-Url: https://codereview.chromium.org/2227343002
Cr-Commit-Position: refs/heads/master@{#411199}