Integrate Crashpad UMA
Creates an global persistent metrics storage helper in base/metrics to
allow stubbing in Crashpad when built outside of Chrome.

Uses a pair of persistent metrics files, one that is the currently
active one, which is renamed to a "previous run" on startup. In normal
operation -active will be created, used, and removed on shutdown.
However, if Crashpad itself crashes, then on next startup, -active will
be moved to previous so that previous can be slurped in on startup to
retrieve data from the crash.

BUG=crashpad:100

Review-Url: https://codereview.chromium.org/2308763002
Cr-Commit-Position: refs/heads/master@{#419251}