Switch base::IDMap to use unique_ptr.
This old base class is a thin wrapper around STL hashmap, providing A)
less easily invalidated iterators and B) a unique_ptr equivalent.  At
this point there is nothing stopping this from *actually* using
unique_ptr, so I switch it over to using that in its IDMapOwnPointer mode.

I made one semantic change, to Replace(), which used to "leak" the
return pointer.  I changed it to instead directly return a unique_ptr
when in that mode, and also to not support replacing an invalid ID.
Both of those changes suit the only user of Replace (in
ServiceWorkerContextCore).

BUG=647091

Review-Url: https://codereview.chromium.org/2340813003
Cr-Commit-Position: refs/heads/master@{#420809}