Cleanup histograms from message loop, as well as kHexRangePrintingFlag.
The message loop histograms are only accessible via command line so are not really that useful. We have similar data via profiler and traces. Note that this logic is very old, added by jar@ from the historical repo.

Removed support for HistogramBase::kHexRangePrintingFlag, which was an available flag for creating histograms. The intention was to make chrome://histograms more readable, but was never really used.

Other minor changes:
 - visual change to chrome://histograms (average -> mean).
 -  made the flag comment a bit more friendly as part of the "make chrome reviews friendlier" effort :)
BUG=651169

Review-Url: https://codereview.chromium.org/2382503002
Cr-Commit-Position: refs/heads/master@{#421650}