Namespace-qualify calls to logging::MakeCheckOpString
Otherwise, argument-dependent lookup will (given that the proper set
of headers are included at the same time) identify WebRTC's
rtc::MakeCheckOpString as an alternative to logging::MakeCheckOpString
if one of the function arguments are from the rtc:: namespace, making
the call ambiguous.

This was the reason why https://codereview.webrtc.org/2384693002 had
to be reverted.

Review-Url: https://codereview.chromium.org/2384263004
Cr-Commit-Position: refs/heads/master@{#422905}