tracing: remove WatchEvent feature
Long time ago WatchEvent introduced the ability to get callbacks for
trace events. This was used by the first versions of the devtools
timeline before it was able to consume the actual trace JSON.
This feature is now unused. Removing it as it causes unnecessary
maintainance burden and is a very dangerous footgun, which allows
a random client to screw up the performance characteristics of tracing.

BUG=

Review-Url: https://chromiumcodereview.appspot.com/2440443002
Cr-Commit-Position: refs/heads/master@{#426233}