Move Java CachedMetrics functionality to base/.
This is general-purpose functionality that allows
recording UMA metrics before the C++ native library
has been loaded and is useful to have outside of the
Chrome layer. In particular, I will be using it from
components/variations to record information about the
first run variations seed fetch.

No functional changes in this CL.

BUG=632199

Review-Url: https://chromiumcodereview.appspot.com/2435813002
Cr-Commit-Position: refs/heads/master@{#426652}