Introduce MemoryCoordinatorProxy in base
This CL instroduces a new class MemoryCoordinatorProxy that is the proxy
of MemoryCoordinator in content/browser. MemoryCoordinator::
GetCurrentMemoryState to get the current memory state is now not
accessible by some MemoryCoordinatorClients, and MemoryCoordinatorProxy
offers the proxy function to access this.

This is used to get the current memory state by e.g. TabLoader in
chrome/browser or components in net.

BUG=639700

Review-Url: https://codereview.chromium.org/2434353003
Cr-Commit-Position: refs/heads/master@{#427027}