Use SharedPersistentMemoryAllocator to share field trial state
Change the method by which we share field trial state from using a
SharedMemory class to SharedPersistentMemoryAllocator. Adds this
allocator to the base::FieldTrialList singleton, so there is only one copy
of this state on the browser process vs. a copy for each process host
which is how it currently works (from
https://codereview.chromium.org/2365273004/)

BUG=653874

Review-Url: https://codereview.chromium.org/2412113002
Cr-Commit-Position: refs/heads/master@{#427378}