Remove CallMsgFilter from MessagePumpWin
kMessageFilterCode is defined locally in message_pump_win.cc.
CallMsgFilter is called with this parameter so that WH_SYSMSGFILTER
and WH_MSGFILTER hooks can look for them.

1) By construction, no hooks can look for kMessageFilterCode since
   there are no hooks in message_pump_win.cc. No other components
   can look for kMessageFilterCode since it is defined locally to
   message_pump_win.cc

2) WH_SYSMSGFILTER and WH_MSGFILTER do not occur anywhere in the
   Chromium codebase.

As a result, we do not need this call in our message pump anymore.

History:
Introduced for dragging a virtual file out of the browser.
https://chromium.googlesource.com/chromium/src/+/0724253eff526440d6262018981438a987f33e69
chrome/browser/views/tab_contents/tab_contents_drag_win.cc

Moved To
* chrome/browser/ui/views/tab_contents/tab_contents_drag_win.cc [1]
* content/browser/tab_contents/web_contents_drag_win.cc [2]
* content/browser/web_contents/web_contents_drag_win.cc [3]

[1] https://chromium.googlesource.com/chromium/src/+/213dac2f0bff9162502fe325b6ebb85a255efcb2
[2] https://chromium.googlesource.com/chromium/src/+/95f072f9ce35f0ecd2cbf899de96fa9448aa1285
[3] https://chromium.googlesource.com/chromium/src/+/0c4e92e6434bb8936aca2e03fe8d44201cb51089

Deleted in
https://chromium.googlesource.com/chromium/src/+/2ceee8f17b6fa286d37f0e9190fed01e5f6a9eff
The associated hook code was not migrated to a different file in this
change.

The WH_MSGFILTER unit test for MessageLoopWin was removed in
https://chromium.googlesource.com/chromium/src/+/6b1f4a6267f234bdbdda1d49f35c04bc3d06795b

BUG=660930

Review-Url: https://codereview.chromium.org/2491403004
Cr-Commit-Position: refs/heads/master@{#431391}