Disallow posting tasks to SequencedWorkerPools by default.
We have observed crash reports in which SequencedWorkerPool workers
were active while redirection to TaskScheduler was enabled. This can
happen if a task is posted to a SequencedWorkerPool before redirection
to TaskScheduler is enabled. With this CL, DumpWithoutCrashing() is
called if a task is posted to a SequencedWorkerPool too early. This
will help us find offending call sites.

A commented DCHECK is also added to find offending call sites. It
will be uncommented in a separate CL to avoid a revert of this CL
in case it fails on the waterfall.

BUG=622400

Committed: https://crrev.com/3ed12bf1d24c19f8af28a62c80c580f8e0d95858
Committed: https://crrev.com/5dfc8358eec8c2ab2a6439bcff115b741093fe2f
Review-Url: https://codereview.chromium.org/2445763002
Cr-Original-Original-Commit-Position: refs/heads/master@{#432913}
Cr-Original-Commit-Position: refs/heads/master@{#433237}
Cr-Commit-Position: refs/heads/master@{#433639}