Add field trial dump and retrieval methods from shared memory
The Breadcrumbs project dumps system state to a persistent file so that it can
be analyzed after a crash.  We want to include in that the complete set of field
trials.  That means saving them to the file when the browser starts and being
able to retrieve them from the file on a future run; they will be loaded into a
protobuf and uploaded.

BUG=663414

Review-Url: https://codereview.chromium.org/2560723004
Cr-Commit-Position: refs/heads/master@{#437732}