PartitionAllocTest: increase address space limit to 6 GB.
In some cases, such as sanitizer_coverage_flags=edge,
more than 5 GB of address space is reserved at a startup.
Raising the address space limit to be higher than that.

Unfortunately, Linux has a broken implementation of getrusage,
and the proper way of fixing this (by setting the current usage + 4 GB)
is not trivial / impossible under a sandbox. This is the riskless
change I could make up: increase the address limit, while still
not increasing the amount of memory allocated to avoid issues on
Android.

BUG=674665

Review-Url: https://codereview.chromium.org/2584123002
Cr-Commit-Position: refs/heads/master@{#439372}