TaskTraits: Rename WithSyncPrimitives() to WithBaseSyncPrimitives().
The goal of this change is to make it clear that there is no need
to try to determine whether external APIs use sync primitives when
choosing TaskTraits. WithBaseSyncPrimitives() should only be used
for tasks that use base/ sync primitives.

In a separate CL, we will make sure that there is a DCHECK failure
whenever a task without WithBaseSyncPrimitives() uses a base/
sync primitive.

TBR=sorin@chromium.org,sdefresne@chromium.org,jam@chromium.org
BUG=675660

Review-Url: https://codereview.chromium.org/2601623002
Cr-Commit-Position: refs/heads/master@{#440652}