Shape detection service: Add QR detection in Mac
This CL lands support for QR/barcode detection in Mac by
adding the correspondent files in services/shape_detection.

Most of the work is reshuffling code that was specific to
Face detection, to suit both Face and Barcode detection:

- Removed shape_detection_service_dispatcher.h and merged
 its two methods as statics in render_process_host_impl.cc.

- Moved common service/ code to detection_utils_mac.h/mm
 (1 helper function)

** Important note re. testing **
Chromium Mac bots seem to have troubles running CoreImage
methods, probably because they are virtual machines (no GPU).
This prevents testing the Barcode detection API properly, and also
renders void the current Face detection test. So, the plan is:
a) Remove the current Mac-Face detection test, because is not
testing anything, really (there was a bug in the error-return path
that is cleaned up here).
b) land (imminently) new unit tests for face and barcode, running
in Mac Gpu bots.

This CL does a); b) will be done in a subsequent CL.

BUG=665150
TEST=
./out/gn/Chromium.app/Contents/MacOS/Chromium --enable-blink-features=ShapeDetection,GeometryInterfaces  https://codepen.io/miguelao/full/wgrYjZ
NOTE: unittests for {face/barcode}_detection_impl_mac.mm coming
imminently after this).

Review-Url: https://codereview.chromium.org/2655303005
Cr-Commit-Position: refs/heads/master@{#447900}