Reland: Avoid calling strlen() where possible in base/trace_config
Fix for crbug.com/706416 was landed and this patch should work fine.

The main target of this CL is to avoid calling strlen() where possible.
It can happen if a raw |char*| literal or result of |c_str()| function is used
to construct a StringPiece or std::string objects.

In addition, StringPiece API like SplitStringPiece was used to avoid
unnecessary string copying and allocations.

BUG=679479
TBR=primiano@chromium.org,yusukes@chromium.org

Review-Url: https://codereview.chromium.org/2780983003
Cr-Commit-Position: refs/heads/master@{#460720}