Remove base::Value dictionary storage indirection.
Previously base::Value used a std::map which was relatively large on Posix and bloated
the size of base::Value by being in the union. Now that base::Value uses a flat_map
which is effectively a std::vector, the size is smaller than std::string and it no
longer bloats the union.

This replaces the unique_ptr indirection with a literl base::flat_map.

Review-Url: https://codereview.chromium.org/2809783002
Cr-Commit-Position: refs/heads/master@{#464778}