memory-infra: Split mojom internal/external response types
Split the ProcessMemoryDump struct into ProcessMemoryDump and
ProcessMemoryDumpInternal. The former is used as part of memory-infra's
public "Request a global memory dump" API and the latter is used for
internal coordination. Spliting them allows us to remove two fields
which were only used in one or other case and also sets the stage for
larger rafactorings to the internal struct to support computing memory
metrics in Chrome.

Bug: 728199
Change-Id: I6603ba4b424bc177082792759be1ace866951828
Reviewed-on: https://chromium-review.googlesource.com/533335
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Primiano Tucci <primiano@chromium.org>
Commit-Queue: Hector Dearman <hjd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#480038}