Show a banner if an unimplemented command-line flag is used.
Some command-line flags (e.g. --enable-heap-profiling) are only
implemented in specific build configurations (e.g. Release+Static), and
in all other binaries we want to warn the user that the flag will be
ignored, if it is specified.

This CL adds a generic "not implemented in this build" banner message
and a special-case to show that message if an --enable heap-profiling
option is specified that is not valid in the current build.

Bug: 705461
Change-Id: Ie41be966ecf17935c52590826901b15cea7171f8
Reviewed-on: https://chromium-review.googlesource.com/517690
Commit-Queue: Wez <wez@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Primiano Tucci <primiano@chromium.org>
Cr-Commit-Position: refs/heads/master@{#480321}