Introduce RunLoop::QuitCurrent*Deprecated() to be favored over MessageLoop::Quit*()
This is required to get rid of direct calls to MessageLoop in tests
and allow for other abstractions to drive RunLoop (since it's impossible
to migrate every test to RunLoop).

Using message_loop_test.cc to test this. Those tests will go away when
the deprecated methods can finally be removed but by then a lot of
callers will be using the RunLoop versions and I don't think it will be
possible to break them.

R=danakj@chromium.org
TBR=dpranke@chromium.org (PRESUBMIT.py)

Bug: 748715
Change-Id: Ia0b996850a47bf24626a99c17b5985e8da826beb
Reviewed-on: https://chromium-review.googlesource.com/585650
Commit-Queue: Gabriel Charette <gab@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#489611}