Replace some uses of std::deque with base::circular_deque in componnets.
This is roughly half of the uses in //components. Some of the uses require
deque and TODOs have been added with the reason why.

#include <algorithm> is added to circular_deque.h for the calls to
std::max. Previous uses apparently included this some other way.

Some nontrivial updates were made to dependency_graph.cc to simplify the
code. In trying to study what this code's requirements were to see if
the usage could be replaced, I realized it could be simplified
by using the (relatively new) base::Erase function. The constructors
that take a range are also used, instead of the existing std::copy with a
back_inserter iterator.

Bug: 757232
Change-Id: I712b1368f5719dfd0bc79421ef3d0d8f58e8239f
Reviewed-on: https://chromium-review.googlesource.com/622887
Reviewed-by: Elliot Glaysher <erg@chromium.org>
Commit-Queue: Brett Wilson <brettw@chromium.org>
Cr-Commit-Position: refs/heads/master@{#496164}