Remove base::PickleSizer
This was added so Mojo IPC could measure the size of an IPC::Message
before serializing it. Mojo IPC doesn't need to do this anymore, so
just delete it and avoid having to maintain a bunch of tests that
Pickle and PickleSizer are in sync.

Change-Id: I67bd8d811aed53f16a7f87f3cce900664a823a94
Reviewed-on: https://chromium-review.googlesource.com/676881
Reviewed-by: Brian White <bcwhite@chromium.org>
Reviewed-by: Ken Rockot <rockot@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#503480}