test: Truncate excessive logs in the middle
Previously we always truncate logs from the beginning, but there could
be important information there. This CL changes the behavior so that we
always truncate the middle part of the log to preserve potentially
important information in the top and bottom part of the log.

BUG=793426
TEST=This improves test infrastructure.

Change-Id: I058f735fbaedec324960b2c7a269d549b1a0178e
Reviewed-on: https://chromium-review.googlesource.com/823191
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#523596}