Support GetOpenFDCount in Android
It is possible to get fd count on Android from current process. Android
does not allow a process to read details about other processes. So,
browser cannot get this data for child. But such is the case for other
function in ProcessMetrics like GetResidentSetSize(). So, it is ok to
enable this function on Android.

BUG=794378

Change-Id: Ib9a139126af131dabba1d3f4ea22cd29c225c8af
Reviewed-on: https://chromium-review.googlesource.com/837904
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Siddhartha S <ssid@chromium.org>
Cr-Commit-Position: refs/heads/master@{#525546}