[Windows] Do not take into account Windows tablet mode when computing interaction MQs.
The interaction media queries are about actual physical input methods.
Windows 8 and 10 allows the user to manually trigger the tablet mode
(settings or taskbar) but it doesn't reflect the actual input
configuration of the device.

This patch modifies isTabletDevice by extracting the code that
is looking at the actual hardware configuration of the device
and the tablet mode. The split is needed because IsTabletDevice
is used by other features in Chrome so I kept the same behavior
(IsTabletDevice is still calling the newly added method). Only the
MQ code is now using the new method which only checks
the hardware configuration to figure out what kind of pointer devices
the machine has.

This will avoid inconsistencies when we detect the hardware change
and the fact that the tablet mode has been switched on/off.

BUG=778845

Change-Id: I13e7d4a2891c38466ed05079fb5d1aa6d4f3e35e
Reviewed-on: https://chromium-review.googlesource.com/809811
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Commit-Queue: Alexis Menard <alexis.menard@intel.com>
Cr-Commit-Position: refs/heads/master@{#528127}