Remove TaskScheduler name from its worker thread's names
A thread is always in the context of a process and adding the process
name to the thread name is redundant (and the name is already long enough
as-is). This keeps TaskScheduler threads with the same name in the
browser and brings other processes on par (just like we don't name the
IO thread differently in other processes).

This is a prerequisite to always require a TaskScheduler name (to always
suffix histograms). Otherwise we would making the browser's worker thread
names even longer.

Also, renamed a few |name| variables to be more descriptive because I
got confused a few times in what each |name| variable held (e.g. PS1).
Intentionally kept |name| in the public interface. That way we can
change how it's used in the impl but having specific names there helps
identify what it's *currently* used for.

Bug: 803019
Change-Id: Iaeadce43706f647a7e8de0f843da35303171a0fa
Reviewed-on: https://chromium-review.googlesource.com/873377
Reviewed-by: François Doray <fdoray@chromium.org>
Commit-Queue: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/master@{#530170}