Build Fuchsia packages for test and executable targets.
All targets that use the test() or fuchsia_executable_runner rules
will produce Fuchsia packages, which contain the test executable and
their runtime dependencies.

The packages are modules which can be deployed in an already-running
system, as opposed to the current scheme which bakes the files into
the system boot image.

Deployment and support for dynamic libraries will be addressed
in followup CLs.

This CL also:
* Fixes PathService logic for packaged invocations.
* Adds a new "build_manifest.py" GN helper script.
* Fixes the dependency graph for Fuchsia runner script generation
  so that it accurately captures the real dependency ordering.


Bug: 796779
Change-Id: I4e3a38e0075573b328ad98425cc493e28cff1ca4
Reviewed-on: https://chromium-review.googlesource.com/841749
Commit-Queue: Kevin Marshall <kmarshall@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Sergey Ulanov <sergeyu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#531288}