Change webview SyncCompositor to use BeginFrameResponse messages.
Instead of sending a sync message we can queue off the fact that we've
sent begin frame requests to the webview. The Webview can deliver the
BeginFrameResponse message.

Move all bookkeeping into a new class SynchronousCompositorSyncCallBridge
which is a shared object between IO/UI threads instead of all of it
being inside the SynchronousCompositorBrowserFilter. This allows easier
adoption of the mojo API which is coming.

BUG=722928

Change-Id: I415ea68f9b258d118905b54f9c17de685cd15dce
Reviewed-on: https://chromium-review.googlesource.com/883563
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#533458}