Reland: Make interactive tests have unlimited timeout on Windows.
This upgrades --interactive to a test-launcher-scope switch (necessarily
renaming it to --test-launcher-interactive) and treats it as a timeout
specifier.  For browser UI tests on Windows, this means the parent process will
now get an (effectively) unlimited timeout and not just the child process, so
the test won't terminate early.

BUG=none
TEST=none

Change-Id: Ic94c5bff322846c2b7f69cad4b805f549dbcaa29
Reviewed-on: https://chromium-review.googlesource.com/919074
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Trent Apted <tapted@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#537211}