Support periodic collection in the ThreadProfiler
Implements support for periodic collection of stack samples following
the initial startup profiling. This change effectively enables periodic
profiling for the GPU process main and IO threads and the renderer
compositor thread.

The periodic thread profiling support currently in use in the browser
process will be removed in a follow-on CL, and the UI and IO thread
users will be updated to use this implementation.


Bug: 808588
Change-Id: I42dc32a7898f03dda0b0419f16171bf6f5d1b1c7
Reviewed-on: https://chromium-review.googlesource.com/907703
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Commit-Queue: Mike Wittman <wittman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#538010}