Fix ThreadLocalStorageTest,UseTLSDuringDestruction to work on Fuchsia.
The original version of the test used base::SimpleThread, which uses Chrome TLS
behind the scenes. This ends up tickling a difference between Fuchsia and other
POSIX implementations with respect to the order in which pthread keys are
destroyed.

The new version of the test uses pthread APIs to avoid this, which fixes the
bug.

Change-Id: I8eaf3230a31f96f9ee03afa88a1b7ee8b770bb42
BUG: 828897, 825218
Reviewed-on: https://chromium-review.googlesource.com/996038
Reviewed-by: Erik Chen <erikchen@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Commit-Queue: Erik Chen <erikchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#548204}