VR: Ensure Chrome launches onto the primary display for VR Standlones
There's a platform bug with Standalone VR devices where transitioning
between the virtual and primary display causes crashes in the
framework. This means that currently (M67) Chrome crashes 100% of the
time on launch, regardless of how Chrome was launched, on Standalones.

I'm not 100% clear on how moving between the displays works on
standalones as it's complicated. Here are the things I know:
1. 2D apps by default launch to the virtual display.
2. Toggling VR mode can cause your Activity to be moved to the primary,
real, display.
3. Moving between displays is causing platform crashes due to the
display being cached in the View Root, and probably elsewhere.

The recommended workaround by the platform folks is to just make sure
Chrome is never launched onto the Virtual Display. This has some
implications. The main implication is that the Chrome window when using
the 2D-in-VR path is super tall as they have to keep text readable with
the app rendering at real display size. When we ship browsing on
standalones, you'll never see the tall window anyways, so it won't be
an issue in the future.

I would provide screenshots, but that's impossible :)

Bug: 838358
Change-Id: I618a01e9e020dec2abcc8f1e65ce7a24842c2113
Reviewed-on: https://chromium-review.googlesource.com/1036117
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/master@{#555040}