Tracing: Add a metadata event to track chrome library start address
On Android for trace symbolization the start address of the library
cannot be obtained by offsets from memory maps since it is mapped from
apk. So, add explicit metadata for each process to mark the start
address of the executable. The only library mapped from apk is
chrome.so.

Change-Id: Ib6f3ec81fd1c34c92974fce7a73debd4f9e8dad5
Reviewed-on: https://chromium-review.googlesource.com/1068239
Commit-Queue: Siddhartha S <ssid@chromium.org>
Reviewed-by: oysteine <oysteine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#560777}