Fix flake in layout tests when running with network service.
The problem was that there were alive processes when tests finished, and swarming considered that a failure (e.g. [1]).

The cause was that the processes were using a job object (to fix other races), and the code in base launched processes suspended if they were in a job object and later resumed them. The race was that layout test script kills the browser when there are any test failures. It did this using "taskkill.exe /t" (recursive). Windows would delete the network process, and as it was killing the browser it would sometimes do that in between the small window where the relaunched network process was created suspended and it was resumed. Child processes have logic to kill themselves if the parent is gone, but since the child process was suspended that didn't run.

The suspension when a process is in a job object was added in r100027 [2]. However the code that was moved there (test server) didn't have this originally. Also, the documentation that describes that this is needed [3] is for functionality that ended in Windows XP [4].

Note that a few weeks ago relaunching started using this as well in [5]. I'm removing created-as-suspended for that code as well, since that seems rare and keeping code in base/ that has this race condition that leaves processes around seems much worse than the alternative race of windows sometimes not being in foreground.

[1] https://chromium-swarm.appspot.com/task?id=3dd10a30220aa810&refresh=10&show_raw=1
[2] https://codereview.chromium.org/7789018
[3] https://msdn.microsoft.com/en-us/library/windows/desktop/ms681949(v=vs.85).aspx
[4] https://msdn.microsoft.com/en-us/library/windows/desktop/ms684159(v=vs.85).aspx
[5] https://chromium-review.googlesource.com/c/chromium/src/+/1064510

Bug: 820996
Change-Id: Ie79118efffac347bf3df1d65e1124fba095acdc0
Reviewed-on: https://chromium-review.googlesource.com/1081559
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Commit-Queue: John Abd-El-Malek <jam@chromium.org>
Cr-Commit-Position: refs/heads/master@{#563967}