TaskScheduler: Run background tasks in foreground pools on platforms without background thread priority.
On platforms that don't support background thread priority, having separate pools
to run background tasks is useless. In fact, it can allow background tasks to run
before foreground tasks (e.g. if the foreground pool is flooded and the background
pool is idle).

This CL removes the background pools on platforms that don't support background
thread priority, and simply runs tasks of all priorities in the same pools, in
order of priority. The size that the background pools would have had is used as
a cap for the number of background tasks that can run in the foreground pools.
This ensures that a foreground pool can't be flooded with background tasks (which
could increase the latency of future foreground tasks).

Bug: 728208
Change-Id: I8fc26e5980acec79b21c084108152220b32bb267
Reviewed-on: https://chromium-review.googlesource.com/1067669
Commit-Queue: François Doray <fdoray@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/master@{#564086}