Add "latency" to list of ETW-traceable events
The "latency" category looks like a perfect event category for ETW
tracing. In particular, TabSwitching::Latency indicates when Chrome
thinks that tab switching has started and stopped which allows it to be
profiled in detail. This has revealed where CPU time is going, and how
long the delay is from when the keyboard event is generated to when
Chrome starts officially processing it. This was helpful in
investigating tab switching latency in bug 858944.

This will require an update to UIforETW to fully enable, which I have
done locally for testing purposes.

Bug: 858944
Change-Id: I4797b804ea6d35458aa3b4e876a78172a29d5131
Reviewed-on: https://chromium-review.googlesource.com/1134516
Reviewed-by: Primiano Tucci <primiano@chromium.org>
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#574586}