Tweak Callback's MaybeValid implementation for smaller binary size
http://crrev.com/bbcf74331e97d451 gained the stripped binary size of
chrome by 40kB (147306728 to 147347688), which is mainly due to an extra
parameter set up on BindStateBase for each base::Bind() caller.

This CL removes the extra argument by merging it to single function with
another parameter. So that the binary loses back to the original size.
(147347688 to 147306728).

Bug: 869098
Change-Id: I7b60797c16a00472e9312466fdd2434395d8bbf7
Reviewed-on: https://chromium-review.googlesource.com/1158093
Commit-Queue: Taiju Tsuiki <tzik@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/master@{#580152}