Fix crash in PEImage::GetDebugId when the pdb info is not present.
The data pointed to by a PE image's debug directory entries need not be
contained in the sections that are mapped into memory. In such cases,
the AddressOfRawData field of the debug directory entry is
null. GetDebugId now handles this by skipping over such entries in its
search for debug info.

Additionally, PEImage::GetDebugId now returns the string length of the
pdb_filename, with proper bounds checking. (Note: StringPiece would have
been a nice fit, but pe_image.{cc,h} does not depend on other parts of
base.)

This change also adds some extra sanity checking to
GetImageDirectoryEntrySize and GetImageDirectoryEntryAddr so that they
return 0/null in case the desired directory entry is not present in the
image. This should never be the case for well-formed images, but we
should be cautious regardless.

BUG=874044
R=brucedawson@chromium.org
TBR=jwd@chromium.org,alph@chromium.org,hjd@chromium.org

Change-Id: I45f8024054c0e3a998dd46bf1d4e9db345138865
Reviewed-on: https://chromium-review.googlesource.com/1175705
Commit-Queue: Greg Thompson <grt@chromium.org>
Reviewed-by: Bruce Dawson <brucedawson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#583860}