Implement signin header modifications for sub-resource requests
This change is a follow-up to r584270 and adds support for modifying
headers sent in sub-resource requests made to the GAIA signon realm.
This is done by inserting a proxy between the render process and network
service when the frame's origin matches the GAIA signon realm.

Bug: 789670
Change-Id: I762a54dd6c8fec57c97f522fc6ef7de2179f3722
Reviewed-on: https://chromium-review.googlesource.com/1187063
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Cr-Commit-Position: refs/heads/master@{#586431}