base: Crash when any shm allocation fails on Mac.
We observe many crashes on Mac caused by shm allocation failures. Our suspision
is that allocations fail because we leak the shared memory, but there is no
signal to help narrow down what that could be. This patch will make the browser
crash as soon as a shm allocation fails, so we can have better stack traces on
the crash server.

Bug: 872237
Change-Id: Ib6bb1a181c0f262a72693e9134061abd707ccc6e
Reviewed-on: https://chromium-review.googlesource.com/1213282
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Alexandr Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#589548}