[Fuchsia] Allow namespace overrides in base::LaunchProcess
Previously base::LaunchProcess didn't allow paths_to_transfer to be used
with the FDIO_SPAWN_CLONE_NAMESPACE flag. This meant that to override
directories the test launcher had to avoid FDIO_SPAWN_CLONE_NAMESPACE
flag. This is not ideal because of the added complexity in the test
launcher and also because the launcher was changing test environment
by not cloning all entries in the namespace (i.e. some tests would behave
differently when started with --single-process-tests flag). This change
updates LaunchProcess() to allow individual dirs transfer while cloning
the rest of the namespace. Also updated test launcher to use this new
feature.

This CL should unbreak the tests currently broken in
https://crrev.com/c/1341243

Change-Id: I3c30497222ffd33d6240f478523d3e767fad90f5
Reviewed-on: https://chromium-review.googlesource.com/c/1345236
Reviewed-by: Scott Graham <scottmg@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#610042}