[scheduler] Making MessageLoop backend variant dependent on Finch config.
We want to A/B test two different backends for MessageLoop: (old) MessageLoopImpl and
(new) SequenceManager, as part of an ongoing effort to get rid of MessageLoop
completely. In order to use Finch, we move its initialisation before that of
MessageLoop in all builds and tests.

Bug: 891670
Change-Id: I808f72b6cf36238f7a2380568a0898ed6c0f5557

Forcing due to fuschia_x64 tests failing because of infrastructure flakiness.
NOTRY=true

Change-Id: I808f72b6cf36238f7a2380568a0898ed6c0f5557
Reviewed-on: https://chromium-review.googlesource.com/c/1333770
Commit-Queue: Karolina Soltys <ksolt@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Alex Clarke <alexclarke@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Cr-Commit-Position: refs/heads/master@{#610876}