[C++ Promises] Add base::TypeId
This is a light weight alternative to RTTI based on using the linker to
uniquely reserve a zero-initialised byte in .rodata. The cost here is
the pointer-sized relocation that the linker must patch in for every
type at program startup. We don't expect vast numbers of types to be
used in practice so this should be OK.

Note we can't use std::any when C++17 arrives because that is based on
RTTI which is banned in Chromium, which motivates base::TypeId.

Bug: 906125
Change-Id: I03f7c14f0cc665d865eb6e029a23f441baf20a9a
Reviewed-on: https://chromium-review.googlesource.com/c/1340317
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Commit-Queue: Alex Clarke <alexclarke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#613486}