Avoid UI work for process launching where possible
In N, a hidden API Context.bindServiceAsUser was added that takes a
Handler parameter that allows the caller to control which thread the
ServiceConnection callbacks happen.
https://android.googlesource.com/platform/frameworks/base/+/691546e5b7f94a6e2d76630ee1287e0b9c69e7a8

For chromium's use cases, using Process.myUserHandle should be the
correct UserHandle. Only need reflection to look up this hidden method,
so reflection is relatively simple. So use this method on N and up.

Note the N implementation did not add a ContextWrapper implementation,
so we need to manually unwrap the Context. It was added to
ContextWrapper in Android O:
https://android.googlesource.com/platform/frameworks/base/+/6192bff1fc3044f6aef5f775f322e640a129ca54

Bug: 901026
Change-Id: Icd97e6774f85d9ad0d4b13af6d0214658d3730dd
Reviewed-on: https://chromium-review.googlesource.com/c/1364294
Commit-Queue: Bo <boliu@chromium.org>
Reviewed-by: ssid <ssid@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Cr-Commit-Position: refs/heads/master@{#614898}