DelayedIncomingQueue::SweepCancelledTasks to operate in place
We've had some OOMs in SweepCancelledTasks because it was hamstrung by
the std::priority_queue API.  Here we take advantage of the fact the
queue is backed by a std::vector and that it's a heap in order to
efficently remove canceled tasks in place.  This should fix OOMs
caused by SweepCancelledTasks.

Bug: 912541
Change-Id: I4bfce9dc399dd90bce90e03e76db38171a575278
Reviewed-on: https://chromium-review.googlesource.com/c/1369780
Commit-Queue: Alex Clarke <alexclarke@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#615377}