Work around Trichrome loading issue.
Android has started loading static shared libraries in a separate
classloader, which means that LibraryLoader can no longer access
non-public fields of NativeLibraries as they are not in the same APK
with the present Trichrome implementation.

We'll be moving to splitting the Trichrome Java code up along package
boundaries when we divide it properly which will avoid this problem, but
this small change works around the problem in the meantime.

Bug: 887078
Change-Id: Ifd0894d4bc2d5a732ffabd37045f2cac70ed6f55
Reviewed-on: https://chromium-review.googlesource.com/c/1376816
Reviewed-by: Eric Stevenson <estevenson@chromium.org>
Commit-Queue: Richard Coles <torne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#616482}