Move TypeId to experimental
TypeId is not quite ready yet. In particular TypeId instances created
for the same type in different components are not "equal".

This patch adds a test that will fail (thus DISABLED for now) and moves
the TypeId class to experimental to prevent people for using it for now.

Bug: 906125
Change-Id: I10f7707d70a66173676a50d7cf3d3a25f431f905
Reviewed-on: https://chromium-review.googlesource.com/c/1377738
Commit-Queue: Gabriel Charette <gab@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/master@{#616787}