Re-enable LogMessageMarkerOnStack test on Windows.
This test was disabled on Windows in
https://chromium-review.googlesource.com/c/1293670
because of failures on the Windows Debug bots.

The failure cannot be reproduced locally, and we don't have links to
logs from the failed runs, so re-enabling the test to allow us to repro
and diagnose the issue.

Bug: 897735, 802393
Change-Id: I9086366e7b1f9240674f3b5678810d221b5d01ff
Reviewed-on: https://chromium-review.googlesource.com/c/1393545
Reviewed-by: Albert J. Wong <ajwong@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Commit-Queue: Wez <wez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#619528}