Add support for ppc64's 64k page sizes in Partition Alloc
Ideally page size would be detected at run-time on ppc64 systems,
but this would most likely require a significant refactor. Since
64k is the de-facto standard and binaries compiled for 64k are likely
to function on 4k systems, it was chosen to use 64k as the default
page size in the partition allocator on ppc64.

This change is part of the effort to upstream a set of patches
that allow chromium to build and run on ppc64le hosts available here:
https://wiki.raptorcs.com/wiki/Porting/Chromium

Change-Id: I23369f261766bdf7aab1b3dc730b94b7f4ba4a96
Reviewed-on: https://chromium-review.googlesource.com/c/1480970
Commit-Queue: Albert J. Wong <ajwong@chromium.org>
Reviewed-by: Albert J. Wong <ajwong@chromium.org>
Cr-Commit-Position: refs/heads/master@{#635426}