[Sampling profiler] Second reland of "Move context frame recording outside the native unwinders"
Allows the native unwinders to be used to unwind from arbitrary frames,
not just from the context frame. This is required to support restarting
native frame unwinding after v8 unwinding.

This is a reland of 866d11f7e6751bd442f7557097afc335cd240360 with a fix
for an MSAN crash.

Bug: 909957, 948302
Change-Id: If0172f8ffac4f056817d32de0c6ef27d403843b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1540285
Commit-Queue: Mike Wittman <wittman@chromium.org>
Reviewed-by: Leonard Grey <lgrey@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#646333}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1549684
Cr-Commit-Position: refs/heads/master@{#647023}