Introduce base/task/promises and DependantList
DependantList is a specalized lock-free list used to record which promises
are dependant on promise. Consider:

    Promise<void> a;

    ...
    Promise<void> b = a.ThenOn(...);
    Promise<void> c = a.ThenOn(...);
    Promise<void> d = a.ThenOn(...);

Promies b, c, & d are dependant on a and will all be in the list. ThenOn
and CatchOn can be called on any thread at any time so we need this to be
thread safe. We only ever need to consume the list once, usually when the
promise is just about to be executed.

Regarding canceled state. This wasn't my original plan but I realised we
need to support some degree of cancellation due to weak pointer
invalidaion. Speficically to make sure we don't leak memory due to
promises that will never be resolved. The semantics are described in the
follow on patch. We'll need to decide if we want to expose this on the
Promise<> template or not.

Bug: 906125
Change-Id: I97e657c24010736a4861a68ef07c3253527c5b8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1539848
Commit-Queue: Alex Clarke <alexclarke@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: Etienne Pierre-Doray <etiennep@chromium.org>
Cr-Commit-Position: refs/heads/master@{#652099}