[fuchsia] Mark VMO from fdio_get_vmo_clone() exec
Fuchsia's fdio_get_vmo_clone is documented as returning a read-only
clone of the underlying VMO, but has long returned a vmo that is both
readable and exec-capable.  As we try to bring the implementation in
line with the docs and make fewer APIs imply executability, we need to
update callers to not rely on the implicit executability of the VMO,
but to add that right explicitly.

Bug: 954263
Change-Id: I4cc14e99b095d432f3163d7d85b70428393e83e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1574688
Commit-Queue: Drew Fisher <zarvox@google.com>
Auto-Submit: Drew Fisher <zarvox@google.com>
Reviewed-by: Wez <wez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#652451}