TaskAnnotator: Remove use of ThreadLocalOwnedPointer.
The overhead introduced by using an owned TLS pointer caused a performance
regression of 0.5% at the median, as this is a hot code path executed for
every single task that is evaluated.

This CL reverts to using a plain TLS pointer as before, and repurposes the
inline |ipc_program_counter| storage in PendingTask to store the current
IPC message context.

BUG=957693

Change-Id: I32ddf9a28b924b5a45301d4855dc08d8b914fbbd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1590247
Commit-Queue: Chris Hamilton <chrisha@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/master@{#656299}