Re-land Adding FastHash() function
This is a reland of e934f4353158fc44f97763aa45380fb449871300 which
was reverted in 34b6010579761fb7e135ec81cc3b050ca23ea867.

The objective is to provide a fast and high quality hash for general
use that will exploit the hardware features of the running CPU.

It is warranted to be high quality and top speed, but it is prone
to be updated when a newer/faster hash function is publicly available.

There are *no warranties* that the hash generated will be the same on
all CPUs (e.g. ARM x Intel) or build target (e.g. 32-bit vs 64-bit).
For such cases, use PersistentHash() instead.

It is also a non-cryptographic hash, therefore, use it when speed
is most important.

For now, it is powered by the latest version of CityHash on both
ARM/Intel and is the same version (i.e. City32() x City64())
depending on now Chromium was built (32-bit app or 64-bit).

Plus migrate server-side *independent* client code to use base::FastHash
that was using deprecated CityHash.

Bug: 902789
Change-Id: Idf92918bd8670db4b07a458940be60e145327e68
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1620931
Commit-Queue: Adenilson Cavalcanti <cavalcantii@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Emil A Eklund <eae@chromium.org>
Cr-Commit-Position: refs/heads/master@{#662367}