Experimentally cap timer interrupt frequency at 125 Hz on DC
Chrome raises the system-global timer interrupt frequency when it thinks
it is necessary in order to ensure responsiveness. The default timer
interrupt frequency is 64 Hz and Chrome raises it to 1 kHz when on AC
power (desktop or plugged in) and 250 Hz when on DC (battery). This is
expressed in terms of the kMinTimerIntervalHighResMs and
kMinTimerIntervalLowResMs constants of 1 and 4 respectively. This CL
changes these constants to functions and adds a disabled-by-default
feature to alter whether MinTimerIntervalLowResMs() should return 8 ms.

Changing to 8 ms will save a noticeable amount of power and therefore
improve battery life. This will be followed by experiments to see
whether this has any undesirable performance side-effects.

This change will have no immediate effect due to the feature being
disabled-by-default.

Bug: 927165
Change-Id: Ibe12b45b0cd9b53929c8ff88a47600fe91d343fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1648820
Reviewed-by: François Doray <fdoray@chromium.org>
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#669326}