[base] Ensure that tests don't change the process priority (reland).
The original CL
https://chromium-review.googlesource.com/c/chromium/src/+/1679434
was reverted because of failures on test end on Mac OS 10.11.
This CL simply disables that check on Mac, so we can benefit from
process priority check on other platforms ASAP.
See diff:
https://chromium-review.googlesource.com/c/chromium/src/+/1754230/1..3/base/test/test_suite.cc

This CL verifies that a test process is not backgrounded when it is
launched and before/after each test. The goal is to avoid having tests
that assume they run at normal priority be disabled because of other
misbehaving tests (e.g. https://crbug.com/931721).

TBR=gab@chromium.org

Bug: 931721
Change-Id: I752121e083370a0e5a6f0e9fb4812de152aa7c51
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1754230
Reviewed-by: François Doray <fdoray@chromium.org>
Commit-Queue: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/master@{#687742}