Mark base::WeakNSObject as test_only
base::WeakNSObject<> is iOS specific code that provides a C++ scoped
object that keeps a weak reference to an Objective-C object. As most
of the iOS specific code has been converted to use ARC, this code is
now only used on code that is shared with macOS and compiled without
ARC.

Move the file from //base to //base/test:test_support so that it can
only be used by testing code (with the goal of eventually removing
it).

Bug: none
Change-Id: Iae30c7adade8dd9caa681881c6e550843f69ed00
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1929228
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Ross McIlroy <rmcilroy@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#718120}