Use base::FastHash() in //base
base::Hash() is deprecated and is being replaced by base::FastHash().
The output of base::FastHash() is only fixed within the lifetime of a
single process, so if the code depends on the mapping of inputs to hash
outputs not changing over time, base::PersistentHash() must be used
instead.

Bug: 1025358
Change-Id: I1b1a46ce59260bb0c9f41a32f8e96f88ff8bf260
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1954454
Auto-Submit: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: kylechar <kylechar@chromium.org>
Reviewed-by: kylechar <kylechar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#722465}