[base] Launch WindowsThreadModeBackground and BoostThreadPriorityOnLibraryLoading features.
With WindowsThreadModeBackground, THREAD_MODE_BACKGROUND_BEGIN is used
to set the priority of background threads. With
BoostThreadPriorityOnLibraryLoading, the priority of a thread is
set to at least normal the first time that a scope in which libraries
are loaded is entered. Stable experiment with these feature show that
they improve input delay and navigation. See experiment results:
https://docs.google.com/document/d/16XrOwuwTwKWdgPbcKKajTmNqtB4Am8TgS9GjbzBYLc0/edit?usp=sharing

This CL enables the features by default. As a safety, we keep the base::Features
in the code until M80 branch.

Bug: 872820
Change-Id: I8894e9fb294220f9daaa2032b258e2a209e78eb8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1945924
Reviewed-by: Brian White <bcwhite@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Commit-Queue: François Doray <fdoray@chromium.org>
Auto-Submit: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/master@{#723203}