[base/task] Remove public API for base::CurrentThread and make destination assert runtime
The destination assert is problematic with the migration to task APIs v3
because they will not require destination to be specified as a
TaskTraits. Keep it as a runtime assert in
post_task.cc:GetTaskExecutorForTraits() to avoid regressions during the
migration.

base::CurrentThread was decided against in task APIs v3:
https://docs.google.com/document/d/1tssusPykvx3g0gvbvU4HxGyn3MjJlIylnsH13-Tv6s4/edit
Removing its public API in this CL and will cleanup the implementation
details when we are more certain exactly which parts of crrev.com/704181
should go.

The Java changes are a clean revert of the additions in
crrev.com/704181 + TabSwitcherMediator.java which was the only user.

base::GetContinuationTaskRunner() is also being reverted but I'm leaving
that one for
https://chromium-review.googlesource.com/c/chromium/src/+/1959917

R=fdoray@chromium.org

Bug: 1026641
Change-Id: I80b616b35be7caec118edb8c51e54633b5743d72
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1978821
Commit-Queue: Gabriel Charette <gab@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Reviewed-by: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730549}