Reformat all gn files in /base/allocator.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=wfh@chromium.org

Bug: 1041419
Change-Id: I1761f9a014f8eff202f780fce5a22f6c02131ffb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998143
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730733}