Reformat all gn files in /base/android/jni_generator/BUILD.gn.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=nyquist@chromium.org

Bug: 1041419
Change-Id: I2b170f09acfa44ada72eb0f88ec1a04e512cfc5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1995791
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Commit-Queue: Tommy Nyquist <nyquist@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730955}