Reformat all gn files in /base.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=gab@chromium.org

Bug: 1041419
Change-Id: I382505127c639e405f4754dee62128afd82c54cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998207
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#731413}