Trace AllowBlocking and AllowIO primitives
This CL is adding trace events for the known issues that
can caused a waitable event or Disk IO on the IO-Thread.
The goal of this change is to allow to track jank and
identify the related bugs by using the slow-reports.

Currently, the ScopedBlockingCall shows a trace event
for the raw primitive but it is time consuming to get back
to the source of the bug. By adding a trace event on the
ScopedAllow*, it is easy to figure out the cause of the
blocking call.

Bug: 1047794
Change-Id: I05734d7344f9eca0307be3af08431fdc5dccf439
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2031440
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Asanka Herath <asanka@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: Joe Downing <joedow@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Jonathan Backer <backer@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/master@{#741464}