Enable memcmp checking in tsan and msan builds.
And fix or suppress a few issues that the msan bot found.

The components/download change is a real bug fix: If we compared N
bytes but the file is only M < N bytes long, we used to read M
bytes and then compare N - M bytes of uninit memory, and if we were
unlucky we could've had a false positive. (Unlikely, but possible.)

Bug: 523428
Change-Id: I01b4c9bf312160eceb81980c731ab301a927341b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2062290
Auto-Submit: Nico Weber <thakis@chromium.org>
Commit-Queue: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742654}