Create trial for HangWatcher.
When the feature is activated HangWatcher will monitor the IOThread
instead of ThreadWatcher. The individual MessageLoop implementations are
instrumented to cover all platforms for the handling of events and
ThreadControllerWithMessagePumpImpl is instrumented to cover the task
execution.

Shutdown hang watching is still covered by ShutdownWatcherHelper.

Bug: 1034046
Change-Id: Ie797a1d1c4ce3b438878ad18e91347ccd02daccf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2046404
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Commit-Queue: Oliver Li <olivierli@google.com>
Cr-Commit-Position: refs/heads/master@{#745143}