gpu: Disable AImageReader for ARM GPUs on Android P and below.
This change adds a temporary workaround to effectively revert the
following changes. While the long term fix for this bug has landed, it
might be too risky to merge to M81. So this change reverts enabling of
AImageReader on these GPUs, to restore the behaviour prior to M80.

Reverted changes:
https://chromium-review.googlesource.com/c/chromium/src/+/2015553
https://chromium-review.googlesource.com/c/chromium/src/+/1873393

TBR=nyquist@chromium.org
R=dalecurtis@chromium.org

Bug: 1051705
Change-Id: I2f7f0828c8ba4a7279c2f6bb95de449402d88f92
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2087987
Commit-Queue: Khushal <khushalsagar@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#747355}