Fix ParameterSet to throw IllegalArgumentException for not supported types.
ParameterSet should throw IllegalArgumentException when we add a
parameter of a type which is not supported in the getAcceptableTypes
method.

However, it throws an ArrayStoreException because the type definition
of ACCEPTABLE_TYPES is of type Set<Class<?>> and when we invoke the
toArray method of the Set class we define the type as String.

This CL fixes that and specifies the correct type which is Class in
the toArray method. It also adds unit tests.

Bug: 1059752
Change-Id: Ia54b0d3acd2e2813a618ad87b97f10a443e64fd4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2093595
Reviewed-by: Benoit L <lizeb@chromium.org>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Commit-Queue: Rohit Agarwal <roagarwal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#748633}