NOTREACHED on unexpected ScopedHString::Create() failure
Normally ScopedHString::Create() should only fail on out-of-memory
errors, which is already handled.

For other errors, currently we use DLOG which will silently pass nullptr
or empty string to downstream code, which may cause other errors which
will be hard to track.

This CL use NOTREACHED() for those errors so we'll get crashes in some
builds if it happens.

Change-Id: Ia7d556d1f981e106fe791c704727f7dd95477c5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2119654
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: Robert Liao <robliao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753374}