Move memory instrumentation service to memory infra thread
The service performs expensive computations on UI thread in both browser
and renderer processes. The computation of VM regions, creating OS
memory dump, graph computation of chrome dumps are all expensive. So,
this moves the coordinator and clients to memory-infra thread which is
created for this purpose.

Change-Id: I6da7e5a0a8d6f8e4b91b8071bf464fb9eaced2ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2109457
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Primiano Tucci <primiano@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Reviewed-by: oysteine <oysteine@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Commit-Queue: ssid <ssid@chromium.org>
Cr-Commit-Position: refs/heads/master@{#753697}