Move Messages for CharacterIndexForPoint to frame.mojom
This CL moves TextInputClientMsg_CharacterIndexForPoint to LocalFrame
and TextInputClientReplyMsg_GotCharacterIndexForPoint to TextInputHost.

It adds GetCharacterIndexAtPoint to frame.mojom with 'is_mac' and
introduces TextInputHost mojom which runs IO thread to get the
response of GetCharacterIndexAtPoint which locks UI thread.

TextInputClientMac is called from RenderWidgetHostViewMac and accesses
to FrameTree to get the focused frame to send GetCharacterIndexAtPoint
to the focused LocalFrame. Since it replaces the legacy IPC with mojo
interface, when it calls GetCharacterIndexAtPoint() it still keeps UI
thread lock/release to mimic Sync IPC on Mac. The result of
GetCharacterIndexAtPoint is passed to TextInputHostImpl which runs IO
thread, it release lock and GetCharacterIndexAtPoint returns the
result got from Blink.

It also updates TextInputClientMacTest to get RenderFrameHost and
stub out local frame mojo binding.

Bug: 1007365
Change-Id: Ic3435d3179410bf0d02bcbbea150d1e0c1b9e9e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2015909
Commit-Queue: Julie Kim <jkim@igalia.com>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761296}