[Fuchsia] Allow ZxHandleWatchController reuse
Previously after aZxHandleWatchController is finished or stopped it was
not possible to re-use for a new handle. That restriction is unnecassary.
This change updates the DCHECK to remove the restriction.

Also replaced mxio->fdio in two stale comments.

Bug: 1077837
Change-Id: I34af04abb9b5e9f6636cf90de40a80bf5d77f017
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2179766
Commit-Queue: Sergey Ulanov <sergeyu@chromium.org>
Auto-Submit: Sergey Ulanov <sergeyu@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765397}