[1/4] Allow registration of power monitor observers pre-initialization
This CL is allowing the registration of PowerObservers before the
PowerMonitor is initialized.

The PowerMonitor has a PowerSource and a ObserverList. The call to
PowerMonitor::Initialize is used to set a PowerSource that will
trigger the notifications. It is thus valid to add observers before
the power source is set.

TBR=miu@chromium.org

Bug: 1074332
Change-Id: Ia238b7caf52f80a7de01329908c4979d81341634
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2202877
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: Jonathan Backer <backer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#771591}