DNR: Log metrics when checksum can't be retrieved for loading ruleset.
Do this by introducing a new entry to the LoadRulesetResult enum which
is logged to metrics. Also, since the enum is now not solely bound to
the RulesetMatcher::CreateVerifiedMatcher result, move it outside the
RulesetMatcher scope and change it to be an enum class.

Also introduce a version of GetBucketCount to HistogramTester which
supports enum classes.

BUG=754526

Change-Id: Ib3a626d4f43581a0e875f3c9b66cc6704b77f550
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2205116
Commit-Queue: Karan Bhatia <karandeepb@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#772070}