base/allocator: Remove SizeSpecificPartitionAllocator.
PartitionAlloc has two variants:
- a general-purpose, thread safe one
- a thread-unsafe, "size specific" one.

The latter only supports "small" allocations, up to a configurable size,
and has a size bucket for each "allocation granularity",
i.e. sizeof(void*). It is currently only used in Blink for layout
objects.

Its use in Blink depends for performance reasons on its absence of
locks, not on the bucketing. Separately, the rest of the thread-safe
variant also supports a thread-unsafe one.

To simplify the code, remove the "size specific" variant, and switch to
the generic allocator everywhere. The advantages are:
- Less code
- Most unit tests were testing the infrequently-used path, not the
  common one
- There were subtle differences between the almost-identical paths in
  each variant.
- Opens the door to further simplidfication, by merging PartitioRootBase
  with its subclass. This will be done in forthcoming CLs.

Note that the partitions are still separate, so this doesn't affect the
security characteristics of the allocator.

Bug: 787153, 998048
Change-Id: I2ccdb6f52ecd1483a0ae84e074c1efdd9422f0f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2235695
Commit-Queue: Benoit L <lizeb@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#777386}