android: Implement bind service fallback logic
This is a general failsafe if the app zygote fails silently, meaning a
bindService call never ends up with a onServiceConnected callback.
Fallback to non-zygote services if onServiceConnected is not called
after a timeout.

Implement by recreating all ChildServiceConnections and making sure old
service connections are never reused. This is to avoid races where
timeout races with the real onServiceConnected call. This is only
enabled on Q+ where onServiceConnected is called on the launcher thread,
not the UI thread, so there is no race from posting to launcher thread.

The switch is static, so after the first timeout, all brand new
connections will immediately fallback to non-zygote.

Bug: 1087966
Change-Id: Iedc2f5c6ddd780c35ca266c257793dbcfdf42e64
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2233312
Reviewed-by: Brian White <bcwhite@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Reviewed-by: ssid <ssid@chromium.org>
Commit-Queue: Bo <boliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779042}