Remove/replace unnecessary includes of logging.h (base/)
If the file just needs the CHECK/CHECK_OP/NOTREACHED
macros, use the appropriate header for that instead.
Or if logging.h is not needed at all, remove it.

This is both a nice cleanup (logging.h is a big header,
and including it unnecessarily has compile-time costs),
and part of the final step towards making logging.h no
longer include check.h and the others.

TBR=thakis

Bug: 1031540
Change-Id: Ib87efde411a3617fcebad6dae08313c20d63d0bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2254284
Commit-Queue: Hans Wennborg <hans@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780580}