[Android] Add utilities for Clank language decouple
This second of four CLs to enable Clank to be in a seperate
locale from the system.

This CL adds the Android preference key APPLICATION_OVERRIDE_LANGUAGE
that will be used to store the override language for Clank.

It also adds utility functions and tests for managing the language
preference and modifying Configuration locales.

Bug: 1068667
Change-Id: Ia8aa6d55c6373696b8af12be43c8741d29bf38d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2355489
Commit-Queue: Trevor  Perrier <perrier@chromium.org>
Reviewed-by: Brandon Wylie <wylieb@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/master@{#802894}