Clean up Safari importing
Starting in macOS 10.14, access to the Safari bookmarks file was
restricted. Chromium didn't notice, and would proceed anyway,
because it assumed that if the file was there, it could be read.
That's not true in the general case anyway, so fix it.

History import from Safari broke many years ago when they moved
to a database for their storage from using a plist. No one realized
that it broke, and since it's access restricted from 10.14 on,
there's no point in fixing it, so this CL removes it.

Bug: 850225
Change-Id: I984ba22a3bcce60ed39cc4fb45331d38f6a6fa94
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2404529
Reviewed-by: Bruce Dawson <brucedawson@chromium.org>
Reviewed-by: Mark Mentovai <mark@chromium.org>
Commit-Queue: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#806257}