[bfcache] Update IPC hash setting functionality in mojo and add unit tests + browser test for IPC tracking to frames in cache
Currently, IPC hash isn't set for tasks posted on the IO thread; adding IPC to SimpleWatcher when tasks are posted allows for the task posted on the IO thread to be annotated with an IPC hash, so that it can then be tracked in the histograms. ScopedSetIpcHash changes are to allow for tracking of IPCs posted from remote mojo objects. However, calculating the MD5 hash of a char* on every posted task is expensive, so we have also introduced storage of the IPC interface name. The IPC interface name can be used to calculate IPC hash on an as-needed basis (i.e. only in cases where IPC is posted to cached frames/documents).

Also includes histogram testing to ensure metrics are recorded. and how ScopedSetIpcHash is set.

Bug: 1110344
Change-Id: I26da160b65145f06639c33679ff03872f686e4af
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2345673
Commit-Queue: Harkiran Bolaria <hbolaria@google.com>
Reviewed-by: François Doray <fdoray@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#807106}