chrome/android: Record radio signal strength and traffic into UMA
Adds a UMA metric to chrome/android that reports signal strength for
Wifi or cellular connection (depending on which is active at the moment)
as well as a number of outgoing and incoming bytes over the last 30 secs.
This metric will be used in conjunction with battery discharge metrics
to estimate the relative importance of radio usage in battery drain.

Bug: b/167544787
Change-Id: I806076896408fd4ac32ca67d21cbb281bea64453
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2450049
Commit-Queue: Mikhail Khokhlov <khokhlov@google.com>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Reviewed-by: Weilun Shi <sweilun@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#815577}