base/tracing: Set async phases for ETW export of typed events
Typed events' phases, categories, and names are already exported to ETW
via TraceLog::ShouldAddAfterUpdatingState(). However, when the typed
event is emitted onto a different track, this wasn't yet reflected in
the phase of the event as exported by ETW.

This patch overrides the phase passed on to ETW export for such events
with explicit tracks.

Bug: 1141860
Change-Id: Ia7b0b5b3dcb44e415110396baf9c2239f2fc656b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2494937
Commit-Queue: Eric Seckler <eseckler@chromium.org>
Reviewed-by: Stephen Nusko <nuskos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#820339}