Hooking up buffer_size in EtwTraceController::StartRealtimeSession
I am working on some ETW stuff for Chrome Remote Desktop and I
noticed that this parameter is unused and a default value of 16KB
is always used.  This isn't a big deal as the only caller appears
to be its unit tests but it seemed like the value passed in
should do something.  The other approach would be to remove the
param but adjusting the buffer size seems like a reasonable thing
to expose (assuming the param is actually used).

Change-Id: I453ac2417d68f6c16fecd517cf1cd74c5aa15731
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2511290
Commit-Queue: Joe Downing <joedow@google.com>
Reviewed-by: Bruce Dawson <brucedawson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#823240}