[ios] Remove incorrect dependency of //base//ios on //ios
Conceptually, //ios is on top of //base, so having a dependency
from //base/ios on //ios is breaking the layering (this also
make it difficult to use //base without using all of Chromium).

Implement IsMultiwindowSupported() and IsSceneStartupSupported()
by checking for the scene API manifest information in the app
Info.plist which is more robust as it will work for any iOS app
build on top of //base that set the information in Info.plist
even if they do not re-use the rest of Chromium.

Bug: none
Change-Id: I5ea006ea617c284dc114b2f8f8ecd6eb2f2dcc1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2529122
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Mike Pinkerton <pinkerton@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#827759}