Create a ResetShortcutsComponent in CCT.
The ResetShortcutsComponent will use the
existing sandboxed shortcut parser to retrieve
modified shortcuts and then overwrites the
existing shortcuts with new ones. The
component will be hooked up to the CCT in a
future CL. The previous CL was previously
reverted due to the target_path FilePath
being equal but the actual string not
equal causing tests to fail. The only difference
between the reverted CL and the new CL is that
the reset_shortcuts_component_unittest now compares
the target_path with the expected path using
PathEquals instead of string comparison of the
FilePath value().
Previous CL:
https://chromium-review.googlesource.com/c/chromium/src/+/2545895

Bug: 1148930
Change-Id: I1a4223632e6ecae6ff8fd0660003f5e4b1d501e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2545650
Reviewed-by: Albert J. Wong <ajwong@chromium.org>
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Commit-Queue: Bettina Dea <bdea@chromium.org>
Cr-Commit-Position: refs/heads/master@{#829506}