base/android: Remove mWaitForSharedRelros from Linker.java
The mWaitForSharedRelros is not used. Deeper code inspection with lizeb@
reveals that it does not cause immediate problems to ignore the state of
this member. For future improvements we should consider sending an empty
bundle with an empty LibInfo and consuming it as usual, instead of
guessing whether to expect this bundle.

Bug: 979638
Change-Id: I1c3369e7bbb818c88da4b64239c918da3c1cda1c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2555120
Auto-Submit: Egor Pasko <pasko@chromium.org>
Commit-Queue: Benoit L <lizeb@chromium.org>
Reviewed-by: Benoit L <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#830552}