Removing bundle from login fails due to errSecParam
As part of the work to support PWA's running at OS login,
I noticed that the call to RemoveFromLoginItems was failing
with errSecParam (-50). The reason appears to be that
the original LSSharedFileListRef is being destroyed, and a new
one is created. This change factors the code so that we
always operate on the same LSSSharedFileListRef which
allows the operation to succeed.

Tested: Manually by mjackson on MacOS 11.0

Bug: 897302
Change-Id: I8d24495b9f3cae3d7df0341a5c5d0cb951e96176
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2522876
Commit-Queue: Mike Jackson <mjackson@microsoft.com>
Reviewed-by: Mark Mentovai <mark@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Reviewed-by: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#832372}