[PartitionAlloc] Don't touch allocator buckets on thread cache hits.
When an allocation hits the thread cache, we still need (at least for
some configurations) to get the allocated size, which is the bucket slot
size. This requires touching PartitionRoot::buckets[bucket_index], which
is not on the same cache line as the rest of the members we access in
PartitionRoot.

Inline the slot size in ThreadCache, to avoid that.

Bug: 998048
Change-Id: Ifad3857b4cadcecd853e7eb0e737a406dc40509f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2566967
Commit-Queue: Benoit L <lizeb@chromium.org>
Reviewed-by: Bartek Nowierski <bartekn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#834709}