sequence_manager: Fix potential crash when sweeping away last task in queue
If the last delayed task in a queue is cancelled and swept away, we
ended up comparing an invalid iterator with the end of the queue.
Depending on where the new end of the queue ends up being, this can
cause us to try to cancel an invalid task, potentially triggering a
crash.

Bug: 1155905
Change-Id: I9df4a2802dbaf3da733f8bde72f394b8fe0b7fab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2612970
Auto-Submit: Sami Kyöstilä <skyostil@chromium.org>
Commit-Queue: Carlos Caballero <carlscab@google.com>
Reviewed-by: Carlos Caballero <carlscab@google.com>
Cr-Commit-Position: refs/heads/master@{#840984}