Replace WMI call in GetHardwareInfoSync with simple registry read
During early startup a call is made to GetHardwareInfoSync which
queries WMI for the model, manufacturer and serial number.  Querying
WMI is an outbound call that can hang.  I have looked at traces
showing 100-500ms outbound hangs.  The model and manufacturer strings
can be read from the registry for windows in a known location.  This
isn't the case for the bios serial number, but it turns out nobody is
using that anyways (at least through GetHardwareInfoSync) so I am
removing it.

Bug: 1164623

Change-Id: Iabb414f42f75008730440529b055d5951ce37dcc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2619419
Reviewed-by: François Doray <fdoray@chromium.org>
Reviewed-by: Owen Min <zmin@chromium.org>
Commit-Queue: Chris Davis <chrdavis@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#842638}