[Android] Add memory dump to native filling controller
The manual filling controller is responsible to forward all
filling data (autofill, passwords, addresses, payments) to
the Android UI.
To do that quickly, it currently uses caches which seem to
consume significant memory. The possibility to record heap
dump helps to put a number on that cost and measure the
effectiveness of our memory savings iterations.

Bug: 1169167
Change-Id: I2b7633e7121f72d548d8d394694f91ff78066806
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2624653
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Matthias Körber <koerber@google.com>
Reviewed-by: ssid <ssid@chromium.org>
Cr-Commit-Position: refs/heads/master@{#847104}