Change FileDescriptorWatcher::Controller::watcher_ to raw pointer
Watcher::WillDestroyCurrentMessageLoop() does "delete this" while it is owned by a unique_ptr in Controller.
This creates a state where unique_ptr is a dangling pointer.
This is weird and also causes issues when CheckedPtr rewrite is applied.
This CL changes FileDescriptorWatcher::Controller::watcher_ to raw pointer

Change-Id: Ia0c38dd0a61297e3d9926c5b322bf066784b8422
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2678242
Commit-Queue: Keishi Hattori <keishi@chromium.org>
Reviewed-by: Bartek Nowierski <bartekn@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#851424}