base/allocator: Properly route UncheckedMalloc to UncheckedAlloc
Makes UncheckedMalloc call UncheckedAlloc so that it returns
nullptr in case of allocation failures.

Existing "#if OS_ANDROID" branch in UncheckedMalloc shouldn't
have been so helpful because the underlying allocator (
PartitionAlloc or the system default) is fixed at build time.
So, this patch removes the #if branch and lets UncheckedMalloc
call UncheckedAlloc appropriately.

Bug: 1111332
Change-Id: I53e8d5ae940da746df8beeafeae80b5634245e56
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2693889
Reviewed-by: Benoit L <lizeb@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Yuki Shiino <yukishiino@chromium.org>
Cr-Commit-Position: refs/heads/master@{#854160}