Ignore UNINSTANTIATED_PARAMETERIZED_TEST errors for TaskRunnerTest
This problem was discovered by attempting to roll googletest past CL
315255779, which causes test binaries to fail when they include test
suites / parametrized tests that are not instantiated, and when they
include empty test suites.

Since this is a template, we do not actually want to run these tests.
This CL suppresses these errors.

Bug: 1163396
Change-Id: I0cd62e1d8d4ab432978304e2ce86d2ebb35f3de0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2688563
Reviewed-by: Victor Costan <pwnall@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Commit-Queue: Austin Sullivan <asully@chromium.org>
Cr-Commit-Position: refs/heads/master@{#854428}