[PA] Finchify lazy commit
There is a suspicion that lazy commit may have caused regression that is
more visible Win 32 than Win 64. This adds a Finch flag to test this.

Note that it isn't possible to disable lazy commit right from the
beginning, because Finch flags can't be read during PartitionAlloc
initialization. This setting is re-configured in main().

Bug: 1181311
Change-Id: Ib24260056d015c20baf673d0ba17e3195327e06d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2696990
Commit-Queue: Bartek Nowierski <bartekn@chromium.org>
Reviewed-by: Anton Bikineev <bikineev@chromium.org>
Reviewed-by: Benoit L <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#857626}