PCScan: Allocate mojo::Channel on non-scannable partition
The CL introduces a non-scannable partition, the purpose of which is
to exclude objects that are known to not contain pointers and thereby
speed up the scanning.

mojo::Channel::Message is currently the only type, allocation of which
is moved to the new partition. All mojo messages amount to roughly 4-5%
of the overall browser memory (tested in cnn:2020).

Bug: 1129751
Change-Id: I0159f7c01d36174cba54f710a82aa86672862efc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2678949
Commit-Queue: Anton Bikineev <bikineev@chromium.org>
Reviewed-by: Bartek Nowierski <bartekn@chromium.org>
Reviewed-by: Benoit L <lizeb@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#859600}