[Base] Properly stop child processes in MultiprocessTestClientLauncher
MultiprocessTestClientLauncher#waitForMainToReturn() isn't closing the
connection after waiting for main to exit. This leads to
ChildConnectionAllocator.java potentially exhausting its connections.

This was discovered when attempting to make sharding aware of batched
android tests[1]. That change resulted in ipc_tests ending up with a
different test order triggering this bug.

When the child process main exits,
ChildProcessConnection.ServiceCallback#onChildProcessDied should
probably be called, but the only non-test callers of this are
SpareChildConnection which is not relevant in this case and
ChildProcessConnection#notifyChildProcessDied().

notifyChildProcessDied() is only called outside of tests if
- ChildProcessConnection start() fails
- ChildProcessConnection#kill() is called
- ChildProcessConnection#stop() is called

ChildProcessConnection#kill() is only called outside of tests in
content's ChildProcessLauncherHelperImpl.

ChildProcessConnection#stop() is called outside of tests by
- ChildProcessLauncher#stop() via content's
  ChildProcessLauncherHelperImpl on shutdown or when a crash is
  detected.
- ChildServiceConnectionImpl#onServiceDisconnected() which is actually
  only invoked inside other tests.

This is a non-issue outside of MultiprocessTestClientLauncher since
all other ChildProcessLauncher invocations go through
ChildProcessLauncherHelperImpl which works correctly. The content/
approach to calling stop relies on IPC detecting a disconnect or
intentional closure. However, MultiprocessTestClientLauncher is used
to test IPC so it is probably best to not rely on IPC detecting an
issue to cleanup.

Instead, MultiprocessTestClientLauncher#waitForMainToReturn() is always
called during test cleanup so simply adding this one stop() invocation
if the ChildProcessLauncher is still around (i.e. start() didn't fail)
should cover all cases.

Please let me know if I've overlooked something or there is a better
solution as I'm not very familiar with this code.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/2725041

Change-Id: I28314889ee156689ae0c165bbcd30acc7ff12cda
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2738603
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/master@{#860405}