ClipboardAndroid::IsFormatAvailable should not access clipboard data
On the native side, ClipboardAndroid holds a cache of the system
clipboard. The cache is not only a copy the system clipboard, but also
holds some formats which Android does not supported.

Before this CL, ClipboardAndroid updates the cache by
UpdateFromAndroidClipboard() in the following three methods.
  HasFormat()
  GetFormats()
  Get()
HasFormat() checks if the format is available in the clipboard, and
GetFormats() checks what formats are available in the clipboard.

In this CL, UpdateFromAndroidClipboard() is removed from HasFormat() and
GetFormats() since they do not need real data to check the content.
Instead check with real data, they just check the mime type in the
clipboard.

This CL also update the way to clear the clipboard by a new API for
Android P+.

Bug:1194592

Change-Id: I859a2225686eceffc6c6c47e159335ebb6de61b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2795984
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Darwin Huang <huangdarwin@chromium.org>
Reviewed-by: Filip Gorski <fgorski@chromium.org>
Commit-Queue: Gang Wu <gangwu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#869778}