[base tracing] Move message pump trace events to disabled-by-default-base
These events are useful in some traces but are so high frequency that
they fill up the trace buffer too much for generic scenarios.

The only event kept is DispatchMessage as this is the native equivalent
of RunTask. Moved it to after ProcessPumpReplacementMessage() to avoid
a double-event when the first ProcessMessageHelper() processes
kMsgHaveWork and recursively invokes itself.

R=etienneb@chromium.org, fdoray@chromium.org

Bug: 899897
Change-Id: I6895789fcbb15308892a9ab9f8e9ac62ba7b617f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2845112
Commit-Queue: Gabriel Charette <gab@chromium.org>
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Etienne Bergeron <etienneb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#875468}