Memlog: Merge samples with identical stack traces.
Prior to this CL, Memlog uploads 1 HeapCallstack per sample. Given that
samples are already a poisson sampling of the underlying allocations, it
doesn't make sense to have separate HeapCallstacks. This means that when
a given stack trace is responsible for a large number of allocations,
the allocations require further post-processing to be grouped together.

This CL merges the HeapCallstacks based on the stack trace, and sums the
total and counts.

Bug: 1202305
Change-Id: Ic2d00258c8bece09ea54d76852b47f21fbad3b08
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2845197
Reviewed-by: ssid <ssid@chromium.org>
Commit-Queue: Erik Chen <erikchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#877515}