Fix shutdown hang in ChromeJsErrorReportProcessor
On ChromeOS, the error processor spawns crash_reporter and then waits
for the subprocess to finish. This can take 10 seconds or more if the
crash_reporter needs to gather lots of logs (or the machine is just
overloaded). During this time, the Chrome task was just waiting.

If, during this time, the user logs out or shuts down the machine, the
task doesn't exit because there's no way to cancel a background task
that's waiting for something else. This leads to Chrome having a
shutdown hang.

Change the approach to launch the process and then schedule another task
to clean up a minute later. If shutdown occurs, the second task will
just not run.

This has two negative effects:
1. It increases the odds that the temporary file will be left around;
before, it was only left behind if crash_reporter was still running when
Chrome was SIGKILLed, so there was only a few second window when the
shutdown request had to happen. Now, if the shutdown request happens
during a full minute, the temp file is left behind. But it's in /tmp so
it will still eventually be cleaned up; plus, as most boards have
memfd_create implemented, they no longer use /tmp files.

2. The time before the completion callback is called increases from
around 5 to 10 seconds normally to 1 minute, which means that for
Essential Apps/Inputs, the crashReportPrivate call isn't resolved for
longer.

But that's better than a shutdown hang.

Bug: chromium:1206022

Change-Id: Icb523a5afa100fc6a76a78651da5dd8546975391
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2893622
Commit-Queue: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Darren Shen <shend@chromium.org>
Cr-Commit-Position: refs/heads/master@{#883763}