Added multiple callback support for histograms in StatisticsRecorder
This CL adds multiple callback support for histogram samples. The
callback to be invoked is wrapped inside a ScopedHistogramSampleObserver
which unregisters itself on destruction. A base::ObserverListThreadSafe
is used for each histogram to maintain the list of observers.

There is a slight change in the behavior of these observers in that
the samples are now always posted, whereas before they were notified
synchronously on the sequence the histogram was recorded on. The
posting of samples happens in the same thread in which the observer
was registered.

Bug: 1119851
Change-Id: I6369d418d9318d22d3927c6d7dd723dd2210f22a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2896125
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Reviewed-by: ssid <ssid@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Weilun Shi <sweilun@chromium.org>
Commit-Queue: Shakti Sahu <shaktisahu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#889590}