[PartitionAlloc] Record the bucket size when freelist corruption is detected.
Freelist corruption happens in the wild, but is very hard to debug,
partly because it means that another object with the same size is likely
being used after being freed. Record the allocation size at crash time,
so that we can at least narrow it down to fewer objects, and see whether
trends appear.

This should be doable with stack traces, but with inlining they are not
very informative.

In terms of hot path code, this adds a cache-hot move. From an assembly
diff on Linux:

 11d:	4b 8b 44 25 00       	mov    0x0(%r13,%r12,1),%rax
 122:	48 85 c0             	test   %rax,%rax
 125:	0f 84 a1 02 00 00    	je     3cc <base::PartitionRoot<true>::AllocFlagsNoHooks(int, unsigned long, unsigned long)+0x3cc>
 12b:	49 83 85 a8 05 00 00 	addq   $0x1,0x5a8(%r13)
 132:	01

 133:	43 0f b7 7c 25 0a    	movzwl 0xa(%r13,%r12,1),%edi    <---- THIS IS NEW

 139:	48 8b 10             	mov    (%rax),%rdx
 13c:	48 85 d2             	test   %rdx,%rdx
 13f:	74 18                	je     159 <base::PartitionRoot<true>::AllocFlagsNoHooks(int, unsigned long, unsigned long)+0x159>
 141:	48 89 d1             	mov    %rdx,%rcx
 144:	48 f7 d1             	not    %rcx
 147:	48 39 48 08          	cmp    %rcx,0x8(%rax)
 14b:	0f 85 b5 02 00 00    	jne    406 <base::PartitionRoot<true>::AllocFlagsNoHooks(int, unsigned long, unsigned long)+0x406>
 151:	48 0f ca             	bswap  %rdx
 154:	0f 18 0a             	prefetcht0 (%rdx)
 157:	eb 03                	jmp    15c <base::PartitionRoot<true>::AllocFlagsNoHooks(int, unsigned long, unsigned long)+0x15c>

Which is touching almost the same location as
    11d:	4b 8b 44 25 00       	mov    0x0(%r13,%r12,1),%rax
above, and given alignment, should not add a cache miss.

Bug: 998048
Change-Id: I1eaa45cd96bd4591a56fc97aee7a14b6bc178cf9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2933356
Reviewed-by: Bartek Nowierski <bartekn@chromium.org>
Commit-Queue: Benoit L <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#890727}