base/android: Teach Linker to find the webview memory reservation
This change is mostly about what the title says, with unittests, but not
only.

Most importantly, on the level of the Linker and its subclasses this
change should add enough support to allow the LibraryLoader/etc to
create RELRO regions both in the browser and in the app zygote (the
latter starting from Android S).

In particular, to simplify the upper layers sometimes the RELRO bundle
will arrive as null, teach the takeSharedRelrosFromBundle() to ignore
these until something usable arrives.

The full end-to-end change doing all the communication is not yet
polished, and not rebased on top of this one yet, sorry. Providing it
for reference:
   https://chromium-review.googlesource.com/c/chromium/src/+/2752872

The rest does a few minor cleanups, renames, rewrites for stale
comments, etc. Could not resist.

Bug: 1154224
Change-Id: Ia5512dbb5d19dab375108d0703994920bba7a3d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2980352
Commit-Queue: Egor Pasko <pasko@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/master@{#901242}