[PartitionAlloc] Try to avoid major page faults in locked sections.
At free(ptr) time, PartitionAlloc writes to *ptr to link its
freelist. If *ptr is very cold, and has been moved to swap, this causes
a major page fault, since the data needs to be brought back from swap.

When this happens in a low priority thread, and the allocation size is
too large to be handled by the thread cache, this write occurs while
holding the partition lock. Since there is no priority inheritance in
our locks, if the OS scheduler doesn't properly boost the blocked
process priority, we can end up with a severe priority inversion.

This happens on Windows (see linked bug), where it is responsible for
(in particular) shutdown hangs. To mitigate that, touch the data
*before* taking the lock.

Bug: 1228523
Change-Id: I7ddf22a9bdf027264fe7040627f6e88caa9b92a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3030942
Reviewed-by: Bartek Nowierski <bartekn@chromium.org>
Commit-Queue: Benoit L <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#902004}