PrintBackendServiceManager uses base::ScopedAllowBlocking
Crash keys support was added to PrintBackendServiceManager in [1].
This has started crashing on Print Preview invocation, with the
backtrace showing that it is marked as blocking from a scope which does
not allow blocking.

Since this usage is invoked from the Browser's UI thread, need to
update this to use base::ScopedAllowBlocking.

[1] https://crrev.com/899887

Bug: 809738, 1230218
Change-Id: I4c3ae6c9eb23349e06653ef20ac71bd397d1e203
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3036624
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Alan Screen <awscreen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#903167}