Enable browser UI stretch on overscroll
Match the behavior of content overscroll by using shared
ElasticOverscroll feature. Because we need to apply near activity
creation, this needs to be cached, as native is not loaded yet.
Unfortunately the content reads this in renderer processes and the
browser cannot easily share its cached version. This will lead to
temporarily inconsistencies during the first restart after a config
change.

Also adjusted content elastic overscroll on Android to only be on S+
instead of R+. This lets the browser and content area match each other,
as native Android stretch on overscroll is only available on S+.

Bug: 1217189
Change-Id: I9c6b147e9ee7536c43ff953de5659ac762fb970c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3026578
Commit-Queue: Sky Malice <skym@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Reviewed-by: Henrique Nakashima <hnakashima@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Reviewed-by: Peter Boström <pbos@chromium.org>
Reviewed-by: Wenyu Fu <wenyufu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#903568}