[PartitionAlloc] Revert PartitionAlloc-Everywhere on Chrome OS.
This is meant to allow comparison between PA-E and the previous
allocator. As such, this CL will be reverted.

Note for performance sheriff: This is going to change a *lot* of
metrics, if there is a memory regression in range, this is the culprit,
most likely.

Bug: 1236604
Change-Id: I5ce1302e383356a862b3630a4a22d755b03d3463
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3071442
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Commit-Queue: Benoit L <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909753}