Extend BarrierCallback to be usable with any type T.
This CL teaches BarrierCallback how to handle Ts that are references and
may be const; and how to work with `done_callback`s that accept a
`[const] std::vector<T>[&]`, instead of just a `std::vector<T>`. This is
helpful for working with Mojo callbacks, since those callbacks are
passed by const ref.

Change-Id: Ifb71dcefa226f76e0d000d828fb5ef18002189f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3085890
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Chris Fredrickson <cfredric@chromium.org>
Cr-Commit-Position: refs/heads/master@{#910582}