Fewer defines in windows_types.h
Now that windows.h is included in fewer Chromium translation units (from
a peak of 70+% to less than 10% for the Chromium repo) it is not
necessary to define so many windows.h macros in windows_types.h.

More macros could be defined if we scattered undef statements around
(see crrev.com/c/3123125 for some examples) but that is not clearly
better.

Bug: 796644
Change-Id: I282a42115bb05cf99761acbb7b7494291830eb45
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3140022
Reviewed-by: Emily Shack <emshack@chromium.org>
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#918286}