Attempting to fix CalendarFactory ANR
This is a speculative attempt to fix a longstanding ANR. We hope that by
timing out a background task attempting Calendar.getInstance(), we can
call it on the main thread and get the result before we ANR.

Note - this may not work. If Calendar.getInstance() is just very slow,
there might not be anything we can do.

Bug: 1241398
Change-Id: I575241f9fd7028ab985d08f23ed13383a3271fa2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3129922
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Commit-Queue: Sam Maier <smaier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#919252}