Replace DISALLOW_COPY_AND_ASSIGN in base/
This replaces DISALLOW_COPY_AND_ASSIGN with explicit constructor deletes
where a local script is able to detect its insertion place (~Foo() is
public => insert before this line).

This logic is wrong if ~Foo() has comments above it, as it should be
inserted before that.

This is incomplete as not all classes have a public ~Foo() declared, so
not all DISALLOW_COPY_AND_ASSIGN occurrences are replaced.

IWYU cleanup is left as a separate pass that is easier when these macros
go away.

Bug: 1010217
Change-Id: Ia83e0ea9132ef47450b22a6b4331a242e7519e66
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3163384
Reviewed-by: Lei Zhang <thestig@chromium.org>
Owners-Override: Lei Zhang <thestig@chromium.org>
Commit-Queue: Peter Boström <pbos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#921905}