[PartitionAlloc] Count dirty memory caused by empty SlotSpans.
Empty slot spans stay temporarily in the per-PartitionRoot ring buffer,
which is costing memory. This CL adds accounting to report the amount of
dirty memory cached in these empty slot spans. It is only reported in
the pa_tcache_inspect tool, as this is a dynamic property, and due to
periodic purge, may be mostly 0 in periodic memory-infra dumps.

We don't report committed memory (which would include the whole slot
span), but dirty one, as one of the aims of this CL is to investigate
why increasing the size of the ring increased private memory footprint.

And thanks to this instrumentation, we find that the ring may not be
purged for a long time in renderers, even though
PartitionAllocMemoryReclaimer should take care of it. This is likely due
to idle tasks (somewhat counter-intuitively) stopping when a renderer is
idle.

The instrumentation is useful beyond this investigation though, since we
likely will want to both have periodic reclaim and cap the total size of
memory wasted this way.

Bug: 1250288
Change-Id: Id64f95a218c82488e79a8a91320cb05c655a2f8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3178988
Commit-Queue: Benoit L <lizeb@chromium.org>
Reviewed-by: Bartek Nowierski <bartekn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#924710}