Add a temporary header for each file moving to base/task/
A multitude of task-related headers currently live in the general
base/ folder. To improve the organisation of the files in the base
library, those files will be moved to base/task/.

Those temporary headers, whose purpose is to redirect the include to
the real location of the header, will facilitate the mass-migration of
all the references to the moved files.

Bug: 1255932
Change-Id: I397021ae29a8042d6ef1b336404864d08497752f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3209027
Auto-Submit: Patrick Monette <pmonette@chromium.org>
Commit-Queue: Gabriel Charette <gab@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/main@{#929305}