Make PaymentRequest a DocumentService
The PaymentRequest's lifetime is bounded by
a) Its mojo receiver's connection
b) The current document in the RenderFrameHost that spawned it
c) It self-destroys on an error, which acts as a shortcut for a).

We would like to:
- avoid holding GlobalFrameRoutingIds, and doing null-checks,
- avoid the fact that the mojo connection can slightly outlive
  the RenderFrameHost, and
- avoid tracking the fragile-and-ever-changing WebContentsObserver
  API to manage the object's lifetime.

DocumentService will ensure that the PaymentRequest (as its
subclass) is destroyed appropriately, and never has a null, or
invalid freed pointer back to the RenderFrameHost.

This change allows us to remove the whole abstraction of the
PaymentRequestWebContentsManager. The PaymentRequest's lifetime
is now fully self-managed, and the object can just be created
directly.

R=creis@chromium.org, rouslan@chromium.org

Bug: 1251691, 1252046
Change-Id: I80ed0ede531df2c177b0246eba83a6abe034cd6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3198719
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#929879}