[fuchsia] Implement fuchsia.process.lifecycle.Lifecycle
base::ProcessLifecycle implements the Lifecycle service for the current
process, invoking a supplied OnceClosure if/when the framework requests
that the process gracefully teardown.

Components be opted-in to receiving graceful teardown events, in order
for this to be used.

Bug: 1242886, 1250747
Change-Id: Ibaf54a788500e47ef34fba1028a82cd312cb5125
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3168783
Commit-Queue: Wez <wez@chromium.org>
Commit-Queue: Benjamin Lerman <qsr@chromium.org>
Auto-Submit: Wez <wez@chromium.org>
Reviewed-by: Benjamin Lerman <qsr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#930560}