[fuchsia][android][ios] Expect failure for unsupported BasePathKeys
Previously, the PathServiceTest.Get test skipped PathService keys that
are not supported on a specific platform. This CL changes the test to
expect failure in those cases. This also ensures that the test gets
updated if the behavior changes.

The updated test caught that DIR_HOME home is actually supported on
Android despite the test previously saying that it is not implemented.

Also add not-implemented logging for DIR_USER_DESKTOP and DIR_HOME on
Fuchsia.

Bug: 1231928, 1257402
Change-Id: Ifc67f3da6b318eb8d093f299d6ee7f785276c5ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3213101
Auto-Submit: David Dorwin <ddorwin@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Commit-Queue: David Dorwin <ddorwin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#933767}