Add WIN_ prefix to Windows-specific base::File flags.
This brings documentation about the flaghs being Windows-only to
the callsites, making the behavior clearer to the calling code.

This is the initial CL that only adds aliases for these flags and
updates references from base/files. The plan is to then send
individual CLs (via git cl split) to migrate call sites, so that
callers have a chance to review their usage for correctness.

Bug: 1244149
Change-Id: I61f63ad5c0a5fff989f3cdc73f67c0f942134c5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3227343
Reviewed-by: Gabriel Charette <gab@chromium.org>
Commit-Queue: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/main@{#933879}