[PartitionAlloc] Improve debugging in crash reports.
In several places, we use base::Alias() to make sure that a given
variable is pushed to the stack. This then in turns allows to see data
in minidumps. But it is not always easy to find the data there, leading
to hacks such as surrounding the data with "sentinel" values, as in
PartitionFreelistEntry for instance.

This CL changes that, by putting a key and a value together on the stack.
For instance, with:
PA_DEBUG_DATA_ON_STACK(1, "size", 0x42)

The values on the stack are (where the bounds are %rsp and %rbp,
respectively):
(lldb) x 0x00007fffffffd0d0 0x00007fffffffd0f0
0x7fffffffd0d0: 73 69 7a 65 00 00 00 00 42 00 00 00 00 00 00 00
size............

Bug: 998048
Change-Id: I845d3d17d49a1ceac794712b87ef877934e7b713
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3240949
Reviewed-by: Bartek Nowierski <bartekn@chromium.org>
Commit-Queue: Benoit L <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#934861}