base: Add TimeDelta::is_positive() and TimeDelta::is_negative()
Currently to check the sign of a TimeDelta, one has to compare it with
TimeDelta(), with the prior knowledge that TimeDelta() default to zero.
This is a bit hard to write and read.

This CL introduces TimeDelta::is_positive() and TimeDelta::is_negative()
to make the check easier and more explicit.

Change-Id: I1ea6bc108efe33ced3554524e1edfffcbe9497a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3238792
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Owners-Override: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/main@{#935233}