[PartitionAlloc] Sort the SlotSpan freelists in PurgeMemory().
To improve cache efficiency and possibly reduce fragmentation as well,
slots to be allocated should be sorted. However this is currently not
the case, as we prefer a Last In First Out (LIFO) ordering, for better
cache locality (since a recently freed allocation is likely still hot in
cache).

These two types of cache locality are not contradictory: allocated
memory from the central allocator should be as contiguous as possible,
and preferably sorted in ascending order (so that newly-allocated memory
looks like a contiguous array), while reused memory should prefer the
most recently used memory. So we want:
- contiguous allocations from the central allocator
- LIFO ordering in the thread cache.

This CL adds a way to sort freelists, but doing it infrequently enough
to not undermine cache locality, hence it is done in PurgeMemory(). Note
that this is not enough, as thread cache batch fill is still
reversed. Also, the end of slot spans are not decommitted, contrary to
what is done for larger bucket sizes (when individual allocations are at
least one page long). This will be explored in forthcoming CLs.

Change-Id: Icb93269b6e1eacbefecfd736b7f97bada883c45a
Bug: 998048
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3172645
Reviewed-by: Bartek Nowierski <bartekn@chromium.org>
Commit-Queue: Benoit L <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#935432}