Add DelayedTaskHandle
This class will be use in a future CL to provide an alternative to
PostDelayedTask which returns a handle to the posted task which can be
used to cancel it afterwards.

DelayedTaskHandle is created with a delegate parameter that is
responsible for the cancellation of the task. This way, different
SequencedTaskRunner implementations can behave differently depending on
their needs. That delegate is ref-counted so that the task queue
and the handle can both own the handle delegate at the same time.

This CL also adds a default implementation for the
DelayedTaskHandle::Delegate abstract class which works by cancelling the
posted task using weak pointers.

Bug: 1262205
Change-Id: Iba8927e8ce8b91252bac96bb86dcb1106120e13f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3035586
Commit-Queue: Patrick Monette <pmonette@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/main@{#936417}