[PartitionAlloc] Defensive crashing in PartitionFree() on Linux.
When PartitionFree() is passed a pointer it did not allocate, the
allocator will most likely crash at a random place. Make sure we crash
at a well-defined one.

This is intended to help debugging crbug.com/1266412, which could be
caused by this type of problem, even though this should never happen on
Linux.

Bug: 1266412
Change-Id: I941a797a2748e6d27c614a92832dbf9d32c793ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3268919
Commit-Queue: Benoit L <lizeb@chromium.org>
Reviewed-by: Bartek Nowierski <bartekn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#940336}