[PartitionAlloc] Allow double zone registration on mac for app_shim.
Since the main binary of the app shim depends on //base, it registers
its own malloc zone. Later, when the main framework is loaded, the code
in InitializeDefaultMallocZoneWithPartitionAlloc() will find that there
is already a zone names "PartitionAlloc", and not register a new one.

However, base::UncheckedMalloc() will still go to the library instance
of PartitionAlloc, as base::UncheckedMalloc() cannot call malloc()
directly, since we configure it to crash rather than return nullptr. If
a pointer which has been base::UncheckedMalloc()-ed in the library gets
free()-ed, then the system free() does not know the zone it comes from,
and panics.

The ideal solution is to not depend on //base in the shim. A workaround
is to allow the library to also register a zone.

The second approach is taken in this CL: we change the zone name so that
the library also registers it. We could remove the check for unicity in
InitializeDefaultMallocZoneWithPartitionAlloc(), but we want to
discourage multiple loading, so isolating the workaround makes more sense.

Bug: 1274236
Change-Id: I8f1f7833b016ad877e520d0f9435cfd4661460ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3322973
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Reviewed-by: ccameron chromium <ccameron@chromium.org>
Commit-Queue: Benoit Lize <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#950976}