[base] Add base::UncheckedFree() to free() memory from UncheckedMalloc()
Chromium configures malloc() to crash rather than returning nullptr, as
some callers are not prepared to handle nullptr properly, and crashing is
preferable to introducing a security issue. In some cases though,
base::UncheckedMalloc() is used to explicitly ask for a nullptr return
value.

However, it is implicitly assumed that a pointer allocated this way can
be passed to free(). This is not the case on all platforms, in all
configurations. For instance, on macOS, there are (rare) cases where
base::UncheckedMalloc() is PartitionAlloc, while free() isn't
necessarily aware of PartitionAlloc. This leads to crashes, which can be
hard to diagnose since they are relatively rare.

To mitigate that, introduce base::UncheckedFree(). This is not enforced,
yet, as call site conversion is split into another CL.

Bug: 1274236, 1279371
Change-Id: Ib10d1cc17b8856dd5af03e0724c35b2423a43e28
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3329862
Reviewed-by: Wez <wez@chromium.org>
Commit-Queue: Benoit Lize <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#951002}