EnumSet: crash various methods when given out-of-range enum values
After uploading the `FromEnumBitmaskIgnoresExtraBits` test that
exercises `ToEnumBitmask()` and `FromEnumBitmask()`, danakj@ pointed out
that silently ignoring out-of-range enum values may be incorrect.

It dawned on me how confusing that is when I saw that a test like the
following passes:

````
  enum class SomeEnum {
    Min,
    Foo,
    Max,
    AfterMax,
  };
  using SomeEnumSet = EnumSet<SomeEnum, SomeEnum::Min, SomeEnum::Max>;

  SomeEnumSet enums(SomeEnum::Foo, SomeEnum::AfterMax);
  SomeEnumSet reconstructed;
  for (SomeEnum val : enums)
    reconstructed.Put(val);
  EXPECT_EQ(enums, reconstructed);
````

The above test only passes because `SomeEnum::AfterMax` was ignored by
the constructor. Technically, `enums` and `reconstructed` are equal, but
neither contains `AfterMax`!

With this CL, the following methods crash when they receive out-of-range
enum values:
  * The variadic constructor
  * single_val_bitstring()
  * Put()
  * FromRange()

I'm intentionally leaving Remove() and Has() as no-ops when the input is
out of range. It seems reasonable that a caller may want to explicitly
check that their EnumSet does not contain Foo, even though Foo is out of
range.

Change-Id: I4c46059df206a789bfccd1bd7ab89aa8ca21a682
Bug: 1278383, 1278484
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3327788
Reviewed-by: danakj chromium <danakj@chromium.org>
Commit-Queue: Dan McArdle <dmcardle@chromium.org>
Cr-Commit-Position: refs/heads/main@{#952026}