Add LockScreenService and implementation
This adds the mojo interface LockScreenService which will allow a
renderer to implement the lock screen API described at
https://github.com/WICG/lock-screen. A test is added with it.

Data is stored in a ValueStore in a similar way to the Chrome Apps
lockscreen API, as described in:
https://docs.google.com/document/d/1zy2cHMHF0gIdVXW0QWE2bX6Zfkw2h5QV5cBepl-B7L8/edit#heading=h.fwni50g0985m
I2P thread: I2P: https://groups.google.com/a/chromium.org/g/blink-dev/c/006tkVwj1yw

The data gets stored outside of the user's profile directory so that the
lock screen profile can access it as described here:
https://docs.google.com/document/d/1ocWgI51LZ3cjsTjVPgHlM2ymR6rwTceRlmYK0c8pGmo/edit#heading=h.bhrn8mszcke3
We use a hash of the origin as the directory for storing the data.

Code to encrypt this data will be added in a followup.

Bug: 1268227
Change-Id: Ief08511dbe12d3b7badf14289b2ddb88378596fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3271971
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Reviewed-by: Glen Robertson <glenrob@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Raymes Khoury <raymes@chromium.org>
Cr-Commit-Position: refs/heads/main@{#952823}