base: Do not wrap TRAP_SEQUENCE() in lambda in IMMEDIATE_CRASH() with clang
The lambda was added to work around a gcc issue in
https://chromium-review.googlesource.com/c/chromium/src/+/960861/8/base/logging.h
with a check that didn't apply it for clang.

That conditional got lost in
https://chromium-review.googlesource.com/c/chromium/src/+/1298091,
probably accidentally.

Put it back.

The lambda is inlined so the generated _code_ doesn't change, but the
inlining still generates debug information, so this helps the compiler
since it needs to produce less debug info. Also, the crash server is
starting to show stack frames for inlined call frames, and this gets
rid of the inline call frame for that lambda.

No intended behavior change.

Bug: 1284379
Change-Id: I0020479e0cb7ee2d6ad6b89c2e2af4a1b342f6da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3370038
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: danakj chromium <danakj@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#956278}