DanglingPtr: Fix lazily_deallocated_deque
This is the #5 source of dangling raw_ptr: 1.4% of the occurrences,
when running ContentShell.

Memory is owned by this class, which would make std::unique_ptr<T[]> the
immediate solution, but not. It instead allocate all the memory and use
placement new on it later. Instead, this patch just reorders
dereferencing and releasing the memory to avoid the raw_ptr to be
temporarily dangling.

This avoid going through the path of putting the memory region under
quarantine and the cost of memsetting the memory on free.

See doc:
https://docs.google.com/document/d/15qq2wr7Xza5awgJAfXUzO1UZFAG7lXDm4K8hd7vC3GU/edit#heading=h.p5mua6e7y11x

Dangling ptr preliminary list:
https://docs.google.com/spreadsheets/d/1BAfz1uHhMkFe2sK2vyTHsJ733VQUhY9v2M4HxOwaM7E/edit?resourcekey=0-H9SwkJTMbBxQPKUEuIrwow#gid=0

See also:
https://docs.google.com/document/d/1fZN-7def91IgXUexhg-8bu2F6Lo7YrXcy0v3rbjZRNg/edit

Bug: 1291138
Change-Id: Iaeb8d1ec8d846e5857fe525afbf5a7c22e1288f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3417214
Reviewed-by: Bartek Nowierski <bartekn@chromium.org>
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Commit-Queue: Alexander Timin <altimin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#967838}