[PartitionAlloc] Increase the pool size on macOS
Some pages can allocate more than 8GiB in the malloc() partition. This
was allowed on macOS previously, as there is no sandbox memory limit
there. The move to PartitionAlloc thus led to some pages crashing due to
OOM which were not previously.

Increase the GigaCage size on macOS as a consequence. While a more
general and better change is in
https://chromium-review.googlesource.com/c/chromium/src/+/3383838, it is
larger (depends on another change), and as a result harder to
cherry-pick.

Bug: 1232567
Change-Id: Ic785f77b896b56682fbd5c5effd28a4462de3b3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3447347
Reviewed-by: Bartek Nowierski <bartekn@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Benoit Lize <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#968357}