Rephrase overview to clarify raw_ptrs need initialization
I had misread the overview paragraph and got the impression that raw_ptr members now no longer need explicit initialization, having missed the part that this only applies on platforms where USE_BACKUP_REF_PTR is enabled. I think this is a bit of a trap, and it would be useful to more clearly state that explicit initialization is still needed for consistent behavior.

Change-Id: I75a7ee149e830642fce3ca7f88bdaa3af2c90861
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3462155
Reviewed-by: Bartek Nowierski <bartekn@chromium.org>
Commit-Queue: Klaus Weidner <klausw@chromium.org>
Cr-Commit-Position: refs/heads/main@{#971287}